AISCSteel

Documentation for AISCSteel.

AISCSteel.classify_section_for_lb_case1

AISCSteel.classify_section_for_lb_case10

AISCSteel.classify_section_for_lb_case13

AISCSteel.classify_section_for_lb_case15

AISCSteel.flexure_capacity_f2

AISCSteel.flexure_capacity_f3

AISCSteel.flexure_capacity_f4

AISCSteel.flexure_capacity_f5

AISCSteel.import_data

AISCSteel.classify_section_for_lb_case1 — Method

classify_section_for_lb_case1(b, t, E, F_y)

Classifies the section of a member subject to axial compression.

Description of applicable member: Flanges of rolled I-shaped sections, plates projecting from rolled I-shaped sections, outstanding legs of pairs of angles connected with continuous contact, flanges of channels, and flanges of tees.

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

CompressionBucklingType enum whose value is either Nonslender or Slender

Reference

AISC Table B4.1a

AISCSteel.classify_section_for_lb_case10 — Method

classify_section_for_lb_case10(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Flanges of rolled I-shaped sections, channels, and tees.

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.classify_section_for_lb_case13 — Method

classify_section_for_lb_case13(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Flanges of all I-shaped sections and channels bent about their minor axis.

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.classify_section_for_lb_case15 — Method

classify_section_for_lb_case15(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Webs of doubly-symmetric I-shaped sections and channels.

Arguments

h:

For rolled sections: the clear distance between flanges less the fillet at each flange.

For built-up sections: clear distance between flanges when welds are used.

t_w: thickness of the web.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.flexure_capacity_f2 — Function

function flexure_capacity_f2(E, G, F_y, Z_x, S_x, I_y, r_y, h_0, J, C_w, c, r_ts, L_b)

Calculates the moment capacity of the applicable section.

Description of applicable member: Doubly symmetric compact I-shaped members and channels bent about their major axis.

Arguments

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

Z_x: plastic section modulous (inch^4)

S_x: elastic section modulous (inch^3)

r_y: radius of gyration about the y-axis (inch)

h_0: distance between the flange centroids (inch)

J: torsional constant (inch^4)

c:

r_ts:

L_b: unbraced length (inch)

Returns

M_n: moment capacity of the section.

Reference

AISC Section F2

AISCSteel.flexure_capacity_f3 — Method

flexure_capacity_f3(E, F_y, Z_x, S_x, r_y, h_0, J, c, r_ts, L_b, h, t_w, λ, λ_pf, λ_rf, λ_fclass)

Calculates the moment capacity of the applicable section.

Description of applicable member: Doubly symmetric I-shaped members with compact webs and noncompact or slender flanges bent about their major axis.

Arguments

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

Z_x: plastic section modulous (inch^4)

S_x: elastic section modulous (inch^3)

r_y: radius of gyration about the y-axis (inch)

h_0: distance between the flange centroids (inch)

J: torsional constant (inch^4)

c:

r_ts:

L_b: unbraced length (inch)

h: clear distance between flanges less the fillets (inch)

t_w: thickness of the web (inch)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender

Returns

M_n: moment capacity of the section.

Reference

AISC Section F3

AISCSteel.flexure_capacity_f4 — Method

flexure_capacity_f4(E, F_y, Z_x, S_x, r_y, h_0, J, c, r_ts, L_b, h, t_w, λ, λ_pf, λ_rf, λ_fclass)

Calculates the moment capacity of the applicable section.

Description of applicable member: Doubly symmetric I-shaped members with compact webs and noncompact or slender flanges bent about their major axis.

Arguments

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

Z_x: plastic section modulous (inch^4)

S_x: elastic section modulous (inch^3)

r_y: radius of gyration about the y-axis (inch)

h_0: distance between the flange centroids (inch)

J: torsional constant (inch^4)

c:

r_ts:

L_b: unbraced length (inch)

h: clear distance between flanges less the fillets (inch)

t_w: thickness of the web (inch)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender

Returns

M_n: moment capacity of the section.

Reference

AISC Section F4

AISCSteel.flexure_capacity_f5 — Method

flexure_capacity_f5(E, F_y, Z_x, S_x, r_y, h_0, J, c, r_ts, L_b, C_b=1, h, t_w, λ, λ_pf, λ_rf, f_class)

Calculates the moment capacity of the applicable section.

Description of applicable member: Doubly symmetric compact I-shaped members and channels bent about their major axis.

Arguments

E: elastic section modulous.

G: shear modulous.

F_y: yield strength of steel.

Z_x:

S_x:

I_y:

r_y:

h_0:

J:

C_w:

c:

r_ts:

L_b:

C_b:

h:

t_w:

λ:

λ_pf:

λ_rf:

f_class:

Returns

M_n: moment capacity of the section.

Reference

AISC Section F5

AISCSteel.import_data — Method

import_data(lookup_value, lookup_col_name::Symbol, data_location::String)

Searches the specified data location and filters for the first row of data found where the value in the column specified is equal to the lookup value.

Parameters

lookup_value - The value to search for

lookup_col_name::Symbol - the column name to search in

csv_file_path::String - The name of csv file in the data directory