Chapter B - Design Requirements

B4. Member Properties

Classification of Sections for Local Buckling

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case1 — Method

case1(b, t, E, F_y)

Classifies the section of a member subject to axial compression.

Description of applicable member: Flanges of rolled I-shaped sections, plates projecting from rolled I-shaped sections, outstanding legs of pairs of angles connected with continuous contact, flanges of channels, and flanges of tees.

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

CompressionBucklingType enum whose value is either Nonslender or Slender

Reference

AISC Table B4.1a

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case10 — Method

case10(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Flanges of rolled I-shaped sections, channels, and tees.

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case11 — Method

case11(b, t, E, F_L, k_c)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Flanges of built-up I-shaped sections (singly or doubly symmetric).

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_L: nominal compression flange stress above which the inelastic buckling limit states apply (ksi)

k_c:

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case12 — Method

case12(b, t, E, F_L, k_c)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Flanges of built-up I-shaped sections (singly or doubly symmetric).

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case13 — Method

case13(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Flanges of all I-shaped sections and channels bent about their minor axis.

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case14 — Method

case14(d, t, E, F_y)

Classifies the compressive elements of a member subject to flexure.

Description of applicable member: Webs of doubly-symmetric I-shaped sections and channels.

Arguments

d: depth of the WT section

t_w: thickness of the web.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case15 — Method

case15(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Webs of doubly-symmetric I-shaped sections and channels.

Arguments

h:

For rolled sections: the clear distance between flanges less the fillet at each flange.

For built-up sections: clear distance between flanges when welds are used.

t_w: thickness of the web.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case17 — Method

case17(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Webs of doubly-symmetric I-shaped sections and channels.

Arguments

b: width of the flat wall of the flange HSS (inch)

t: thickness of the flange (inch)

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case19 — Method

case19(h, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Webs of doubly-symmetric I-shaped sections and channels.

Arguments

h: height of the flat wall of the web HSS (inch)

t: thickness of the web (inch)

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case20 — Method

case20(D, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Webs of doubly-symmetric I-shaped sections and channels.

Arguments

D: outer diameter of the round HSS (inch)

t: thickness of the wall (inch)

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case3 — Method

case3(b, t, E, F_y)

Classifies the section of a member subject to axial compression.

Description of applicable member: Legs of single angles, legs of double angles with separators, and all other unstiffened elements.

Arguments

b: length of the angle leg.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

CompressionBucklingType enum whose value is either Nonslender or Slender

Reference

AISC Table B4.1a

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case4 — Method

case4(d, t_w, E, F_y)

Classifies the section of a member subject to axial compression.

Description of applicable member: Flanges of rolled I-shaped sections, plates projecting from rolled I-shaped sections, outstanding legs of pairs of angles connected with continuous contact, flanges of channels, and flanges of tees.

Arguments

h: depth of WTShape.

t_w: thickness of the web.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

CompressionBucklingType enum whose value is either Nonslender or Slender

Reference

AISC Table B4.1a

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case5 — Method

case5(h, t_w, E, F_y)

Classifies the section of a member subject to axial compression.

Description of applicable member: Flanges of rolled I-shaped sections, plates projecting from rolled I-shaped sections, outstanding legs of pairs of angles connected with continuous contact, flanges of channels, and flanges of tees.

Arguments

h:

For rolled sections: the clear distance between flanges less the fillet at each flange.

For built-up sections: clear distance between flanges when welds are used.

t_w: thickness of the web.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

CompressionBucklingType enum whose value is either Nonslender or Slender

Reference

AISC Table B4.1a

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case6 — Method

case6(b, t, E, F_y)

Classifies the section of a member subject to axial compression.

Description of applicable member: Walls of rectangular HSS

Arguments

b: width of the flat wall.

t: thickness of the wall.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

CompressionBucklingType enum whose value is either Nonslender or Slender

Reference

AISC Table B4.1a