Chapter E - Design of Members for Compression

This chapter applies to members subject to axial compression.

E1. General Provisions

E2. Effective Length

E3. Flexural Buckling of Members without Slender Elements

E4. Torsional and Flexural-Torsional Buckling of Single Angles and Members without Slender Elements

E5. Single Angle Compression Members

E6. Built-Up Members

E7. Members with Slender Elements

E1. General Provisions

The design compressive strength, ${\phi}_c P_{n}$, and the allowable compressive strength, $P_n/{\Omega}_c$, are determined as follows.

The nominal compressive strength, $P_n$ shall be the lowest value obtained based on the applicable limit states of flexural buckling, torsional buckling, and flexural-torsional buckling.

$ {\phi}_c = 0.90 \text{(LRFD)} $

E2. Effective Length

The effective length, $L_c$, for calculatios of member slenderness, $L_c/r$, shall be determined in accordance with Chapter C or Appendix 7.

E3. Flexural Buckling of Members without Slender Elements

This section applies to nonslender-element compression members, as defined in Section B4.1, for elements in axial compression.

AISCSteel.ChapterECompression.E3 — Module

module E3

This section applies to nonslender element compression members, as defined in Section B4.1, for elements in axial compression.

AISCSteel.ChapterECompression.E3.calc_Fe — Method

calc_Fe(E, L_c, r)

Calculates the elastic buckling stress of the applicable section.

Description of applicable member: member without slender elements.

Arguments

E: modulous of elasticity (ksi)

L_c: effective length of member (inch)

r: radius of gyration (inch)

Returns

F_e: elastic buckling stress (ksi)

Reference

AISC Section E3 (E3-4)

AISCSteel.ChapterECompression.E3.calc_Fn — Method

calc_Fn(L_c, r, F_y, F_e)
calc_Fn(F_y, F_e)

Calculates the nominal stress of the applicable section.

Description of applicable member: member without slender elements.

Arguments

L_c: effective length of member (inch)

r: radius of gyration (inch)

F_y: yield strength of steel (ksi)

F_e: elastic buckling stress (ksi)

Returns

F_n: nominal stress (ksi)

Reference

AISC Section E3

AISCSteel.ChapterECompression.E3.calc_Pn — Method

calc_Pn(F_n, A_g)
calc_Pn(E, L_c, r, F_y, A_g)

Calculates the nominal compressive strength for flexural buckling of the applicable section.

Description of applicable member: member without slender elements.

Arguments

F_n: nominal stress (ksi)

A_g: gross area of member (inch^2)

Returns

P_n: nominal compressive strength for flexural buckling (kip)

Reference

AISC Section E3 (E3-1)

E4. Torsional and Flexural-Torsional Buckling of Single Angles and Members without Slender Elements

This section applies to singly symmetric and unsymmetric members, certain doubly symmetric members, such as cruciform or built-up members, and doubly symmetric members when the torsional unbraced length exceeds the lateral unbraced length, all without slender elements. These provisions also apply to single angles with $b/t > 0.71\sqrt{E/F_y}$ , where $b$ is the width of the longest leg and $t$ is the thickness.

AISCSteel.ChapterECompression.E4 — Module

module E4

This section applies to singly symmetric and unsymmetric members, certain doubly symmetric members, such as cruciform or built-up members, and doubly symmetric members when the torsional unbraced length exceeds the lateral unbraced length, all without slender elements. These provisions also apply to single angles with $b/t > 0.71*sqrt{E/F_y}$

AISCSteel.ChapterECompression.E4.calc_Fex — Method

calc_Fex(E, L_cx, r_x)

Calculates the elastic buckling stress of the applicable section about the x-axis.

Description of applicable member: member without slender elements.

Arguments

E: modulous of elasticity (ksi)

L_cx: effective length of member for buckling about the x-axis (inch)

r_x: radius of gyration about the x-axis (inch)

Returns

F_ex: elastic buckling stress about the x-axis (ksi)

Reference

AISC Section E4 (E4-5)

AISCSteel.ChapterECompression.E4.calc_Fey — Method

calc_Fey(E, L_cy, r_y)

Calculates the elastic buckling stress of the applicable section about the y-axis.

Description of applicable member: member without slender elements.

Arguments

E: modulous of elasticity (ksi)

L_cy: effective length of member for buckling about the y-axis (inch)

r_y: radius of gyration about the y-axis (inch)

Returns

F_ey: elastic buckling stress about the y-axis (ksi)

Reference

AISC Section E4 (E4-6)

AISCSteel.ChapterECompression.E4.calc_Fez — Method

calc_Fez(E, C_w, L_cz, G, J, A_g, r̄_0)

Calculates the elastic buckling stress of the applicable section about the z-axis.

Description of applicable member: member without slender elements.

Arguments

E: modulous of elasticity (ksi)

C_w: warping constant (inch^6)

G: shear modulus elasticity of steel (ksi)

J: torsional constant (inch^4)

A_g: gross area of member (inch^2)

r̄_0: polar radius of gyration about the shear center (inch)

Returns

F_ez: elastic buckling stress about the z-axis (ksi)

Reference

AISC Section E4 (E4-7)

AISCSteel.ChapterECompression.E4.calc_H — Method

calc_H(x_0, y_0, r̄_0)

Calculates the flexural constant.

Description of applicable member: member without slender elements.

Arguments

x_0: x-coordinate of the shear center with respect to the centroid (inch)

y_0: y-coordinate of the shear center with respect to the centroid (inch)

r̄_0: polar radius of gyration about the shear center (inch)

Returns

H: flexural constant

Reference

AISC Section E4 (E4-8)

AISCSteel.ChapterECompression.E4.calc_Pn — Method

calc_Pn(F_n, A_g)

Calculates the nominal compressive strength for flexural buckling of the applicable section.

Description of applicable member: member without slender elements.

Arguments

F_n: nominal stress (ksi)

A_g: gross area of member (inch^2)

Returns

P_n: nominal compressive strength for flexural buckling (kip)

Reference

AISC Section E4 (E4-1)

E5. Single Angle Compression Members

This section is not implemented.

E6. Built-Up Members

This section is not implemented.

E7. Members with Slender Elements

This section applies to slender-element compression members, as defined in Section B4.1 for elements in axial compression.

AISCSteel.ChapterECompression.E7 — Module

module E7

This section applies to slender element compression members, as defined in Section B4.1 for elements in axial compression.

AISCSteel.ChapterECompression.E7.calc_Ae — Method

calc_Ae(A_g, b, b_e, t)

Calculates the effective area for slender elements of the applicable section.

Description of applicable member: member with slender elements.

Arguments

A_g: gross area of member (inch^2)

b width of element (for tees this is d; for webs this is h)

b_e: effective width for slender elements (inch)

t: thickness of element

Returns

A_e: effective area of member (inch^2)

Reference

AISC Section E7

AISCSteel.ChapterECompression.E7.calc_Fel — Method

calc_Fel(c_2, λ_r, λ, F_y)

Calculates the elastic local buckling stress of the applicable section.

Description of applicable member: member with slender elements.

Arguments

c_2: second effective width imprtfection adjustment factor

λ_r: onslender slenderness ratio limit

λ: slenderness ratio

F_y: yield strength of steel (ksi)

Returns

F_el: elastic local buckling stress (ksi)

Reference

AISC Section E7 (F7-5)

AISCSteel.ChapterECompression.E7.calc_Pn — Method

calc_Pn(F_n, A_e)

Calculates the nominal compressive strength for local buckling of the applicable section.

Description of applicable member: member with slender elements.

Arguments

F_n: nominal stress (ksi)

A_e: effective area of member (inch^2)

Returns

P_n: nominal compressive strength for local buckling (kip)

Reference

AISC Section E7 (E7-1)

AISCSteel.ChapterECompression.E7.calc_be — Method

calc_be(λ, λ_r, F_y, F_n, c_1, F_el)

Calculates the effective width for slender elements of the applicable section.

Description of applicable member: member with slender elements.

Arguments

λ: slenderness ratio

λ_r: onslender slenderness ratio limit

F_y: yield strength of steel (ksi)

F_n: nominal stress (ksi)

c_1: effective width imprtfection adjustment factor

F_el: elastic local buckling stress (ksi)

Returns

b_e: effective width for slender elements (inch)

Reference

AISC Section E7 (F7-2 and E7-3)