Chapter F - Design of Members for Flexure

This chapter applies to members subject to simple bending about one principal axis. For simple bending, the member is loaded in a plane parallel to a principal axis that passes through the shear center or is restrained against twisting at load points and supports.

F1. General Provisions

F2. Doubly Symmetric Compact I-Shaped Members and Channels Bent about Their Major Axis

F3. Doubly Symmetric I-Shaped Members with Compact Webs and Noncompact or Slender Flanges Bent about Their Major Axis

F4. Other I-Shaped Members with Compact or Noncompact Webs Bent about Their Major Axis

F5. Doubly Symmetric and Singly Symmetric I-Shaped Members with Slender Webs Bent about Their Major Axis

F6. I-Shaped Members and Channels Bent about Their Minor Axis

F7. Square and Rectangular HSS and Box Sections

F8. Round HSS

F9. Tees and Double Angles Loaded in the Plane of Symmetry

F10. Single Angles

F11. Rectangular Bars and Rounds

F12. Unsymmetrical Shapes

F13. Proportions of Beams and Girders

F1. General Provisions

F2. Doubly Symmetric Compact I-Shaped Members and Channels Bent about Their Major Axis

This section applies to doubly symmetric I-shaped members and channels bent about their major axis, having compact webs and compact flanges as defined in Section B4.1 for flexure.

AISCSteel.ChapterFFlexure.F2.calc_Mn — Function

calc_Mn(E, F_y, Z_x, S_x, r_y, h_0, J, c, r_ts, L_b, C_b=1)

Calculates the moment capacity of the applicable section.

Description of applicable member: Doubly symmetric compact I-shaped members and channels bent about their major axis.

Arguments

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

Z_x: plastic section modulous (inch^4)

S_x: elastic section modulous (inch^3)

r_y: radius of gyration about the y-axis (inch)

h_0: distance between the flange centroids (inch)

J: torsional constant (inch^4)

c:

r_ts:

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_n: moment capacity of the section.

Reference

AISC Section F2

AISCSteel.ChapterFFlexure.F2.calc_MnFY — Method

calc_MnFY(M_p)

Calculates the moment capacity of the applicable section for yielding.

Description of applicable member: Doubly symmetric compact I-shaped members and channels bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

Returns

M_nFY: moment capacity of the section for yielding.

Reference

AISC Section F2.1

AISCSteel.ChapterFFlexure.F2.calc_MnLTB — Function

calc_MnLTB(M_p, F_y, S_x, F_cr, L_b, L_p, L_r, C_b=1)

Calculates the moment capacity of the applicable section for lateral torsional buckling.

Description of applicable member: Doubly symmetric compact I-shaped members and channels bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

F_y: yield strength of steel (ksi)

S_x: elastic section modulous (inch^3)

F_cr: critical stress (ksi)

L_b: unbraced length (inch)

L_p: the limiting laterally unbraced length for the limit state of yielding (inch)

L_r: the limiting laterally unbraced length for the limit state of inelastic lateral-torsional buckling (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_nLTB: moment capacity of the section for lateral torsional buckling.

Reference

AISC Section F2.2

F3. Doubly Symmetric I-Shaped Members with Compact Webs and Noncompact or Slender Flanges Bent about Their Major Axis

This section applies to doubly symmetric I-shaped members bent about their major axis having compact webs and noncompact or slender flanges as defined in Section B4.1 for flexure.

AISCSteel.ChapterFFlexure.F3.calc_Mn — Function

calc_Mn(E, F_y, Z_x, S_x, r_y, h_0, J, c, r_ts, L_b, h, t_w, λ_f, λ_pf, λ_rf, λ_fclass, C_b=1)

Calculates the moment capacity of the applicable section.

Description of applicable member: Doubly symmetric I-shaped members with compact webs and noncompact or slender flanges bent about their major axis.

Arguments

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

Z_x: plastic section modulous (inch^4)

S_x: elastic section modulous (inch^3)

r_y: radius of gyration about the y-axis (inch)

h_0: distance between the flange centroids (inch)

J: torsional constant (inch^4)

c:

r_ts:

L_b: unbraced length (inch)

h: clear distance between flanges less the fillets (inch)

t_w: thickness of the web (inch)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_n: moment capacity of the section.

Reference

AISC Section F3

AISCSteel.ChapterFFlexure.F3.calc_MnCFLB — Method

calc_MnCFLB(M_p, E, F_y, S_x, k_c, λ_f, λ_pf, λ_rf, λ_fclass)

Calculates the moment capacity of the applicable section for compression flange local buckling.

Description of applicable member: Doubly symmetric I-shaped members with compact webs and noncompact or slender flanges bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

S_x: elastic section modulous (inch^3)

k_c:

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

Returns

M_nCFLB: moment capacity of the section for compression flange local buckling.

Reference

AISC Section F3.2

AISCSteel.ChapterFFlexure.F3.calc_MnLTB — Function

calc_MnLTB(M_p, F_y, S_x, F_cr, L_b, L_p, L_r, C_b=1)

Calculates the moment capacity of the applicable section for yielding.

Description of applicable member: Doubly symmetric I-shaped members with compact webs and noncompact or slender flanges bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

F_y: yield strength of steel (ksi)

S_x: elastic section modulous (inch^3)

F_cr: critical stress (ksi)

L_b: unbraced length (inch)

L_p: the limiting laterally unbraced length for the limit state of yielding (inch)

L_r: the limiting laterally unbraced length for the limit state of inelastic lateral-torsional buckling (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_nLTB: moment capacity of the section for lateral torsional buckling.

Reference

AISC Section F3.1

F4. Other I-Shaped Members with Compact or Noncompact Webs Bent about Their Major Axis

This section applies to doubly symmetric I-shaped members bent about their major axis with noncompact webs and singly symmetric I-shaped members with webs attached to the mid-width of the flanges, bent about their major axis, with compact or noncompact webs, as defined in Section B4.1 for flexure.

AISCSteel.ChapterFFlexure.F4.calc_Mn — Method

calc_Mn(E, F_y, Z_x, S_x, S_xc, S_xt, b_fc, t_fc, h, h_c, t_w, J, h_0, I_y, I_yc, λ_w, λ_pw, λ_rw, λ_f, λ_pf, λ_rf, λ_fclass, L_b, C_b=1)

Calculates the moment capacity of the applicable section.

Description of applicable member: Other I-shaped members with compact webs or noncompact webs bent about their major axis.

Arguments

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

Z_x: plastic section modulous (inch^4)

S_x: elastic section modulous (inch^3)

S_xc: elastic section modulous referred to compression flange (inch^3)

S_xt: elastic section modulous referred to tension flange (inch^3)

b_fc: width of compression flange (inch)

t_fc: thickness of compression flange (inch)

h: clear distance between flanges less the fillets (inch)

h_c: twice the distance from the centroid to the following: (inch)

the inside face of the compression flange less the fillet or corner radius for rolled shapes

the nearest line of fasteners at the compression flange or inside face of the compression flange when welds are used for built-up sections

t_w: thickness of the web (inch)

J: torsional constant (inch^4)

h_0: distance between the flange centroids (inch)

I_y: second moment of inertia about the y-axis (inch^4)

I_yc: second moment of inertia about the y-axis for the compression flange (inch^4)

λ_w: slenderness ratio of the flange

λ_pw: compact slenderness ratio limit of the web

λ_rw: noncompact slenderness ratio limit of the web

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_n: moment capacity of the section.

Reference

AISC Section F4

AISCSteel.ChapterFFlexure.F4.calc_MnCFLB — Method

calc_MnCFLB(M_p, R_pc, M_yc, F_L, S_xc, E, k_c, λ_f, λ_pf, λ_rf, λ_fclass)

Calculates the moment capacity of the applicable section for compression flange local buckling.

Description of applicable member: Other I-shaped members with compact webs or noncompact webs bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

R_pc: web plastification factor

M_yc: yield moment in the compression flange

F_L: nominal compression flange stress above which the inelastic buckling limit states apply (ksi)

S_xc: elastic section modulous referred to compression flange (inch^3)

E: modulous of elasticity (ksi)

k_c:

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

Returns

M_nCFLB: moment capacity of the section for compression flange local buckling.

Reference

AISC Section F4.3

AISCSteel.ChapterFFlexure.F4.calc_MnCFY — Method

calc_MnCFY(R_pc, M_yc)

Calculates the moment capacity of the applicable section for compression flange yielding.

Description of applicable member: Other I-shaped members with compact webs or noncompact webs bent about their major axis.

Arguments

R_pc: web plastification factor

M_yc: yield moment in the compression flange

Returns

M_nCFY: moment capacity of the section for compression flange yielding.

Reference

AISC Section F4.1

AISCSteel.ChapterFFlexure.F4.calc_MnLTB — Function

calc_MnLTB(M_p, R_pc, M_yc, F_L, S_xc, F_cr, L_b, L_p, L_r, C_b=1)

Calculates the moment capacity of the applicable section for lateral torsional buckling.

Description of applicable member: Other I-shaped members with compact webs or noncompact webs bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

R_pc: web plastification factor

M_yc: yield moment in the compression flange

F_L: nominal compression flange stress above which the inelastic buckling limit states apply (ksi)

S_xc: elastic section modulous referred to compression flange (inch^3)

F_cr: critical stress (ksi)

L_b: unbraced length (inch)

L_p: the limiting laterally unbraced length for the limit state of yielding (inch)

L_r: the limiting laterally unbraced length for the limit state of inelastic lateral-torsional buckling (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_nLTB: moment capacity of the section for lateral torsional buckling.

Reference

AISC Section F4.2

AISCSteel.ChapterFFlexure.F4.calc_MnTFY — Method

calc_MnTFY(M_p, R_pt, M_yt, S_xc, S_xt)

Calculates the moment capacity of the applicable section for tension flange yielding.

Description of applicable member: Other I-shaped members with compact webs or noncompact webs bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

R_pt: web plastification factor corresponding to the tension flange yielding limit state

M_yt: yield moment in the tension flange

S_xc: elastic section modulous referred to compression flange (inch^3)

S_xt: elastic section modulous referred to tension flange (inch^3)

Returns

M_nTFY: moment capacity of the section for tension flange yielding.

Reference

AISC Section F4.4

F5. Doubly Symmetric and Singly Symmetric I-Shaped Members with Slender Webs Bent about Their Major Axis

This section applies to doubly symmetric and singly symmetric I-shaped members with slender webs attached to the mid-width of the flanges and bent about their major axis as defined in Section B4.1 for flexure.

AISCSteel.ChapterFFlexure.F5.calc_Mn — Function

flexure_capacity_f5(E, F_y, S_xc, S_xt, b_fc, t_fc, h, h_c, t_w, λ_f, λ_pf, λ_rf, λ_fclass, L_b, C_b=1)

Calculates the moment capacity of the applicable section.

Description of applicable member: I-shaped members with slender webs bent about their major axis.

Arguments

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

S_x: elastic section modulous (inch^3)

S_xc: elastic section modulous referred to compression flange (inch^3)

S_xt: elastic section modulous referred to tension flange (inch^3)

b_fc: width of compression flange (inch)

t_fc: thickness of compression flange (inch)

h: clear distance between flanges less the fillets (inch)

h_c: twice the distance from the centroid to the following: (inch)

the inside face of the compression flange less the fillet or corner radius for rolled shapes

the nearest line of fasteners at the compression flange or inside face of the compression flange when welds are used for built-up sections

t_w: thickness of the web (inch)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_n: moment capacity of the section.

Reference

AISC Section F5

AISCSteel.ChapterFFlexure.F5.calc_MnCFLB — Method

calc_MnCFLB(R_pg, F_crCFLB, S_xc)

Calculates the moment capacity of the applicable section for compression flange lateral buckling.

Description of applicable member: I-shaped members with slender webs bent about their major axis.

Arguments

R_pg: bending strength reduction factor

F_crCFLB: critical stress relating to compression flange lateral buckling

S_xc: elastic section modulous referred to compression flange (inch^3)

Returns

M_nCFLB: moment capacity of the section for compression flange lateral buckling.

Reference

AISC Section F5.3

AISCSteel.ChapterFFlexure.F5.calc_MnCFY — Method

calc_MnCFY(M_p)

Calculates the moment capacity of the applicable section for compression flange yielding.

Description of applicable member: I-shaped members with slender webs bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

Returns

M_nCFY: moment capacity of the section for compression flange yielding.

Reference

AISC Section F5.1

AISCSteel.ChapterFFlexure.F5.calc_MnLTB — Method

calc_MnLTB(R_pg, F_crLTB, S_xc)

Calculates the moment capacity of the applicable section for lateral torsional buckling.

Description of applicable member: I-shaped members with slender webs bent about their major axis.

Arguments

R_pg: bending strength reduction factor

F_crLTB: critical stress relating to lateral torsional buckling

S_xc: elastic section modulous referred to compression flange (inch^3)

Returns

M_nLTB: moment capacity of the section for lateral torsional buckling.

Reference

AISC Section F5.2

AISCSteel.ChapterFFlexure.F5.calc_MnTFY — Method

calc_MnTFY(F_y, S_xt)

Calculates the moment capacity of the applicable section for tension flange yielding.

Description of applicable member: I-shaped members with slender webs bent about their major axis.

Arguments

F_y: yield strength of steel (ksi)

S_xt: elastic section modulous referred to tension flange (inch^3)

Returns

M_nTFY: moment capacity of the section for tension flange yielding.

Reference

AISC Section F5.4

F6. I-Shaped Members and Channels Bent about Their Minor Axis

This section is not implemented.

F7. Square and Rectangular HSS and Box Sections

This section is not implemented.

F8. Round HSS

This section is not implemented.

F9. Tees and Double Angles Loaded in the Plane of Symmetry

This section is not implemented.

F10. Single Angles

This section is not implemented.

F11. Rectangular Bars and Rounds

This section is not implemented.

F12. Unsymmetrical Shapes

This section is not implemented.

F13. Proportions of Beams and Girders

This section is not implemented.