CShapes

AISCSteel.Shapes.CShapes.CShape — Type

struct CShape <: AbstractCShapes

CShape in the AISC steel database.

Fields

shape: name of the WShape

weight: weight of section (plf)

area: area of wshape (inch2)

d: depth of wshape (inch)

b_f: width of flange (inch)

t_w: thickness of web (inch)

t_f: thickness of flange (inch)

k: Distance from outer face of flange to web toe of fillet used for design (inch)

h: clear distance between flanges less the fillets (inch)

x:

e_0:

x_p:

I_x: Moment of inertia about the x-axis (inch4)

Z_x: Plastic section modulus about the x-axis (inch3)

S_x: Elastic section modulus about the x-axis (inch3)

r_x: Radius of gyration about the x-axis (inch)

I_y: Moment of inertia about the y-axis (inch4)

Z_y: Plastic section modulus about the y-axis (inch3)

S_y: Elastic section modulus about the y-axis (inch3)

r_y: Radius of gyration about the y-axis (inch)

J: Torsional constant (inch4)

C_w: Warping constant (inch6)

W_no: Normalized warping function, as used in Design Guide 9 (inch2)

S_w1: Warping statical moment at point 1 on cross section, as used in AISC Design Guide 9 (inch4)

S_w2:

S_w3:

Q_f: Statical moment for a point in the flange directly above the vertical edge of the web, as used in AISC Design Guide 9 (inch3)

Q_w: Statical moment for a point at mid-depth of the cross section, as used in AISC Design Guide 9 (inch3)

r_0:

H:

r_ts: Effective radius of gyration (inch)

h_0: Distance between the flange centroids (inch)

PA: Shape perimeter minus one flange surface (or short leg surface for a single angle), as used in Design Guide 19 (inch)

PB: Shape perimeter, as used in AISC Design Guide 19 (inch)

PC: Box perimeter minus one flange surface, as used in Design Guide 19 (inch)

PD: Box perimeter, as used in AISC Design Guide 19 (inch)

T: Distance between web toes od fillets at the top and bottom of web (inch)

WG_i: The workable gage for the inner fastener holes in the flange that provides for entering and tightening clearances and edge distance and spacing requirements. The actual size, combination, and orientation of fastener components should be compared with the geometry of the cross section to ensure compatibility. See AISC Manual Part 1 for additional information (inch)

E: Elastic section modulus (ksi) = 29000ksi

F_y: Yield strength(ksi) = 50ksi

AISCSteel.Shapes.CShapes.CShape — Method

CShape(shape; E=29000ksi, F_y=50ksi, C_b=1)

Constructor for CShape.

AISCSteel.Shapes.CShapes.MCShape — Type

struct MCShape <: AbstractMCShapes

MCShape in the AISC steel database.

Fields

shape: name of the WShape

weight: weight of section (plf)

area: area of wshape (inch2)

d: depth of wshape (inch)

b_f: width of flange (inch)

t_w: thickness of web (inch)

t_f: thickness of flange (inch)

k: Distance from outer face of flange to web toe of fillet used for design (inch)

h: clear distance between flanges less the fillets (inch)

x:

e_0:

x_p:

I_x: Moment of inertia about the x-axis (inch4)

Z_x: Plastic section modulus about the x-axis (inch3)

S_x: Elastic section modulus about the x-axis (inch3)

r_x: Radius of gyration about the x-axis (inch)

I_y: Moment of inertia about the y-axis (inch4)

Z_y: Plastic section modulus about the y-axis (inch3)

S_y: Elastic section modulus about the y-axis (inch3)

r_y: Radius of gyration about the y-axis (inch)

J: Torsional constant (inch4)

C_w: Warping constant (inch6)

W_no: Normalized warping function, as used in Design Guide 9 (inch2)

S_w1: Warping statical moment at point 1 on cross section, as used in AISC Design Guide 9 (inch4)

S_w2:

S_w3:

Q_f: Statical moment for a point in the flange directly above the vertical edge of the web, as used in AISC Design Guide 9 (inch3)

Q_w: Statical moment for a point at mid-depth of the cross section, as used in AISC Design Guide 9 (inch3)

r_0:

H:

r_ts: Effective radius of gyration (inch)

h_0: Distance between the flange centroids (inch)

PA: Shape perimeter minus one flange surface (or short leg surface for a single angle), as used in Design Guide 19 (inch)

PB: Shape perimeter, as used in AISC Design Guide 19 (inch)

PC: Box perimeter minus one flange surface, as used in Design Guide 19 (inch)

PD: Box perimeter, as used in AISC Design Guide 19 (inch)

T: Distance between web toes od fillets at the top and bottom of web (inch)

WG_i: The workable gage for the inner fastener holes in the flange that provides for entering and tightening clearances and edge distance and spacing requirements. The actual size, combination, and orientation of fastener components should be compared with the geometry of the cross section to ensure compatibility. See AISC Manual Part 1 for additional information (inch)

E: Elastic section modulus (ksi) = 29000ksi

F_y: Yield strength(ksi) = 50ksi

AISCSteel.Shapes.CShapes.MCShape — Method

MCShape(shape; E=29000ksi, F_y=50ksi, C_b=1)

Constructor for MCShape.

Compression API for CShapes

This section is not implemented.

Flexure API for CShapes

AISCSteel.Shapes.CShapes.Flexure — Module

module Flexure

This module includes useful functions to calculate bending capacity of rolled C-Shape sections (CShape, MCShape).

Functions

classify_flange_major_axis - classify flange for slenderness when bent about the x-axis

classify_flange_minor_axis - classify flange for slenderness when bent about the y-axis

classify_web - classify web for slnderness

calc_Mnx - moment capacity about the x-axis

calc_Mny - moment capacity about the y-axis

Modules

F2 - includes functions specific for F2 sections

F6 - includes functions specific for F6 sections

AISCSteel.Shapes.CShapes.Flexure.calc_Mnx — Method

calc_Mnx(shape::T, L_b, C_b=1) where T <: AISCSteel.Shapes.CShapes.AbstractCShapes

This function calculates Mnx of the shape.

Arguments

shape: rolled C-Shape section (CShape, MCShape)

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_nx: nominal moment of the section (kip-ft)

Reference

AISC Section F2

AISCSteel.Shapes.CShapes.Flexure.calc_Mny — Method

calc_Mny(shape::T) where T <: AISCSteel.Shapes.CShapes.AbstractCShapes

This function calculates Mny of the shape.

Arguments

shape: rolled C-Shape section (CShape, MCShape)

Returns

M_ny: nominal moment of the section (kip-ft)

Reference

AISC Section F6

AISCSteel.Shapes.CShapes.Flexure.classify_flange_major_axis — Method

classify_flange_major_axis(shape::T) where T <: AISCSteel.Shapes.CShapes.AbstractCShapes

This function classifies flange for flexure for the shape.

Arguments

shape: rolled C-Shape section (CShape, MCShape)

Returns

(λ_f, λ_pf, λ_rf, λ_fclass)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

AISCSteel.Shapes.CShapes.Flexure.classify_flange_minor_axis — Method

classify_flange_minor_axis(shape::T) where T <: AISCSteel.Shapes.CShapes.AbstractCShapes

This function classifies flange for flexure for the shape.

Arguments

shape: rolled C-Shape section (CShape, MCShape)

Returns

(λ_f, λ_pf, λ_rf, λ_fclass)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

AISCSteel.Shapes.CShapes.Flexure.classify_web — Method

classify_web(shape::T) where T <: AISCSteel.Shapes.CShapes.AbstractCShapes

This function classifies web for flexure for the shape.

Arguments

shape: rolled C-Shape section (CShape, MCShape)

Returns

(λ_w, λ_pw, λ_rw, λ_wclass)

λ_w: slenderness ratio of the web

λ_pw: compact slenderness ratio limit of the web

λ_rw: noncompact slenderness ratio limit of the web

λ_wclass: compact noncompact or slender classification for the web