HSS Shapes

AISCSteel.Shapes.HSS_Shapes.HSS_Shape — Type

struct HSS_Shape <: AbstractHSS_Shapes

HSS_Shape in the AISC steel database.

Fields

shape: name of the HSS

weight: weight of section (plf)

A_g: area of shape (inch2)

Ht: overall depth of square HSS or longer wall of rectangular HSS (inch)

h: depth of flat wall of square HSS or longer wall of rectangular HSS (inch)

B: overall width of square HSS or shorter wall of rectangular HSS (inch)

b: width of the flat wall of square HSS or the shorter flat wall of rectangular HSS (inch)

t_nom: nominal thickness of HSS and pipe wall (inch)

t_des: design thickness of HSS and pipe wall (inch)

I_x: Moment of inertia about the x-axis (inch4)

Z_x: Plastic section modulus about the x-axis (inch3)

S_x: Elastic section modulus about the x-axis (inch3)

r_x: Radius of gyration about the x-axis (inch)

I_y: Moment of inertia about the y-axis (inch4)

Z_y: Plastic section modulus about the y-axis (inch3)

S_y: Elastic section modulus about the y-axis (inch3)

r_y: Radius of gyration about the y-axis (inch)

J: Torsional constant (inch4)

C: HSS torsional constant (inch3)

E: Elastic section modulus (ksi) = 29000ksi

F_y: Yield strength(ksi) = 50ksi

AISCSteel.Shapes.HSS_Shapes.HSS_Shape — Method

HSS_Shape(shape; E=29000ksi, F_y=50ksi, C_b=1)

Constructor for HSS_Shape.

Compression API for HSS Shapes

AISCSteel.Shapes.HSS_Shapes.Compression — Module

module Compression

This module includes useful functions to calculate compression capacity of rolled hss-shape sections (HSS_Shape).

Functions

classify_long_wall - classify longer wall for slenderness

classify_short_wall - classify shorter wall for slenderness

calc_Pn - Compressive capacity of the shape

AISCSteel.Shapes.HSS_Shapes.Compression.calc_Pn — Method

calc_Pn(shape::T, L_cx, L_cy) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes
calc_Pn(shape::T, L_cx, L_cy, λ_b, λ_rb, λ_bclass, λ_h, λ_rh, λ_hclass) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes

This function calculates Pn of the shape.

Arguments

shape: rolled hss-shape section (HSS_Shape)

L_cx: effective length of member for buckling about the x-axis (inch)

L_cy: effective length of member for buckling about the y-axis (inch)

λ_b: slenderness ratio of the shorter wall

λ_rb: nonslender slenderness ratio limit of the shorter wall

λ_bclass: nonslender or slender classification for the shorter wall

λ_h: slenderness ratio of the longer wall

λ_rh: nonslender slenderness ratio limit of the longer wall

λ_hclass: nonslender or slender classification for the longer wall

Returns

P_n: nominal compressive strength of the section (kip)

Reference

AISC Section E3, E7

AISCSteel.Shapes.HSS_Shapes.Compression.classify_long_wall — Method

classify_long_wall(shape::T) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes

This function classifies wall for compression for the shape.

Arguments

shape: rolled hss-shape section (HSS_Shape)

Returns

(λ_f, λ_rf, λ_fclass)

λ_f: slenderness ratio of the wall

λ_rf: nonslender slenderness ratio limit of the wall

λ_fclass: nonslender or slender classification for the wall

AISCSteel.Shapes.HSS_Shapes.Compression.classify_short_wall — Method

classify_short_wall(shape::T) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes

This function classifies wall for compression for the shape.

Arguments

shape: rolled hss-shape section (HSS_Shape)

Returns

(λ_f, λ_rf, λ_fclass)

λ_f: slenderness ratio of the wall

λ_rf: nonslender slenderness ratio limit of the wall

λ_fclass: nonslender or slender classification for the wall

Flexure API for HSS Shapes

AISCSteel.Shapes.HSS_Shapes.Flexure — Module

module Flexure

This module includes useful functions to calculate bending capacity of rolled HSS sections (HSS_Shape).

Functions

classify_flange_major_axis - classify flange for slenderness when bent about the x-axis

classify_flange_minor_axis - classify flange for slenderness when bent about the y-axis

classify_web_major_axis - classify web for slnderness when bent about the x-axis

classify_web_minor_axis - classify web for slnderness when bent about the y-axis

calc_Mnx - moment capacity about the x-axis

calc_Mny - moment capacity about the y-axis

AISCSteel.Shapes.HSS_Shapes.Flexure.calc_Mnx — Method

calc_Mnx(shape::T, L_b, C_b=1) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes
calc_Mnx(shape::T, λ_f, λ_pf, λ_rf, λ_fclass, λ_w, λ_pw, λ_rw, λ_wclass, L_b, C_b=1) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes

This function calculates Mnx of the shape.

Arguments

shape: HSS section (HSS_Shape)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

λ_w: slenderness ratio of the web

λ_pw: compact slenderness ratio limit of the web

λ_rw: noncompact slenderness ratio limit of the web

λ_wclass: compact noncompact or slender classification for the web

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_nx: nominal moment of the section (kip-ft)

Reference

AISC Section F7

AISCSteel.Shapes.HSS_Shapes.Flexure.calc_Mny — Method

calc_Mny(shape::T) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes
calc_Mny(shape::T, λ_f, λ_pf, λ_rf, λ_fclass, λ_w, λ_pw, λ_rw, λ_wclass) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes

This function calculates Mny of the shape.

Arguments

shape: HSS section (HSS_Shape)

Returns

M_ny: nominal moment of the section (kip-ft)

Reference

AISC Section F7

AISCSteel.Shapes.HSS_Shapes.Flexure.classify_flange_major_axis — Method

classify_flange_major_axis(shape::T) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes

This function classifies flange for flexure for the shape.

Arguments

shape: rolled HSS sections (HSS_Shape)

Returns

(λ_f, λ_pf, λ_rf, λ_fclass)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

AISCSteel.Shapes.HSS_Shapes.Flexure.classify_flange_minor_axis — Method

classify_flange_minor_axis(shape::T) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes

This function classifies flange for flexure for the shape.

Arguments

shape: rolled HSS sections (HSS_Shape)

Returns

(λ_f, λ_pf, λ_rf, λ_fclass)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

AISCSteel.Shapes.HSS_Shapes.Flexure.classify_web_major_axis — Method

classify_web(shape::T) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes

This function classifies web for flexure for the shape.

Arguments

shape: rolled HSS sections (HSS_Shape)

Returns

(λ_w, λ_pw, λ_rw, λ_wclass)

λ_w: slenderness ratio of the web

λ_pw: compact slenderness ratio limit of the web

λ_rw: noncompact slenderness ratio limit of the web

λ_wclass: compact noncompact or slender classification for the web

AISCSteel.Shapes.HSS_Shapes.Flexure.classify_web_minor_axis — Method

classify_web(shape::T) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes

This function classifies web for flexure for the shape.

Arguments

shape: rolled HSS sections (HSS_Shape)

Returns

(λ_w, λ_pw, λ_rw, λ_wclass)

λ_w: slenderness ratio of the web

λ_pw: compact slenderness ratio limit of the web

λ_rw: noncompact slenderness ratio limit of the web

λ_wclass: compact noncompact or slender classification for the web