IShapes

AISCSteel.Shapes.IShapes.RolledIShapes.HPShape — Type

struct HPShape <: AbstractRolledIShapes

HPShape in the AISC steel database.

Fields

shape: name of the HPShape

weight: weight of section (plf)

area: area of wshape (inch2)

d: depth of wshape (inch)

b_f: width of flange (inch)

t_w: thickness of web (inch)

t_f: thickness of flange (inch)

k: Distance from outer face of flange to web toe of fillet used for design (inch)

k_1: Distance from web center line to flange toe of fillet used for detailing (inch)

h: clear distance between flanges less the fillets (inch)

I_x: Moment of inertia about the x-axis (inch4)

Z_x: Plastic section modulus about the x-axis (inch3)

S_x: Elastic section modulus about the x-axis (inch3)

r_x: Radius of gyration about the x-axis (inch)

I_y: Moment of inertia about the y-axis (inch4)

Z_y: Plastic section modulus about the y-axis (inch3)

S_y: Elastic section modulus about the y-axis (inch3)

r_y: Radius of gyration about the y-axis (inch)

J: Torsional constant (inch4)

C_w: Warping constant (inch6)

W_no: Normalized warping function, as used in Design Guide 9 (inch2)

S_w1: Warping statical moment at point 1 on cross section, as used in AISC Design Guide 9 (inch4)

Q_f: Statical moment for a point in the flange directly above the vertical edge of the web, as used in AISC Design Guide 9 (inch3)

Q_w: Statical moment for a point at mid-depth of the cross section, as used in AISC Design Guide 9 (inch3)

r_ts: Effective radius of gyration (inch)

h_0: Distance between the flange centroids (inch)

PA: Shape perimeter minus one flange surface (or short leg surface for a single angle), as used in Design Guide 19 (inch)

PB: Shape perimeter, as used in AISC Design Guide 19 (inch)

PC: Box perimeter minus one flange surface, as used in Design Guide 19 (inch)

PD: Box perimeter, as used in AISC Design Guide 19 (inch)

T: Distance between web toes od fillets at the top and bottom of web (inch)

WG_i: The workable gage for the inner fastener holes in the flange that provides for entering and tightening clearances and edge distance and spacing requirements. The actual size, combination, and orientation of fastener components should be compared with the geometry of the cross section to ensure compatibility. See AISC Manual Part 1 for additional information (inch)

WG_0: The bolt spacing between inner and outer fastener holes when the workable gage is compatible with four holes across the flange. See AISC Manual Part 1 for additional information (inch)

E: Elastic section modulus (ksi) = 29000ksi

F_y: Yield strength(ksi) = 50ksi

AISCSteel.Shapes.IShapes.RolledIShapes.HPShape — Method

HPShape(shape; E=29000ksi, F_y=50ksi, C_b=1)

Constructor for HPShape.

AISCSteel.Shapes.IShapes.RolledIShapes.MShape — Type

struct MShape <: AbstractRolledIShapes

MShape in the AISC steel database.

Fields

shape: name of the MShape

weight: weight of section (plf)

area: area of wshape (inch2)

d: depth of wshape (inch)

b_f: width of flange (inch)

t_w: thickness of web (inch)

t_f: thickness of flange (inch)

k: Distance from outer face of flange to web toe of fillet used for design (inch)

k_1: Distance from web center line to flange toe of fillet used for detailing (inch)

h: clear distance between flanges less the fillets (inch)

I_x: Moment of inertia about the x-axis (inch4)

Z_x: Plastic section modulus about the x-axis (inch3)

S_x: Elastic section modulus about the x-axis (inch3)

r_x: Radius of gyration about the x-axis (inch)

I_y: Moment of inertia about the y-axis (inch4)

Z_y: Plastic section modulus about the y-axis (inch3)

S_y: Elastic section modulus about the y-axis (inch3)

r_y: Radius of gyration about the y-axis (inch)

J: Torsional constant (inch4)

C_w: Warping constant (inch6)

W_no: Normalized warping function, as used in Design Guide 9 (inch2)

S_w1: Warping statical moment at point 1 on cross section, as used in AISC Design Guide 9 (inch4)

Q_f: Statical moment for a point in the flange directly above the vertical edge of the web, as used in AISC Design Guide 9 (inch3)

Q_w: Statical moment for a point at mid-depth of the cross section, as used in AISC Design Guide 9 (inch3)

r_ts: Effective radius of gyration (inch)

h_0: Distance between the flange centroids (inch)

PA: Shape perimeter minus one flange surface (or short leg surface for a single angle), as used in Design Guide 19 (inch)

PB: Shape perimeter, as used in AISC Design Guide 19 (inch)

PC: Box perimeter minus one flange surface, as used in Design Guide 19 (inch)

PD: Box perimeter, as used in AISC Design Guide 19 (inch)

T: Distance between web toes od fillets at the top and bottom of web (inch)

WG_i: The workable gage for the inner fastener holes in the flange that provides for entering and tightening clearances and edge distance and spacing requirements. The actual size, combination, and orientation of fastener components should be compared with the geometry of the cross section to ensure compatibility. See AISC Manual Part 1 for additional information (inch)

WG_0: The bolt spacing between inner and outer fastener holes when the workable gage is compatible with four holes across the flange. See AISC Manual Part 1 for additional information (inch)

E: Elastic section modulus (ksi) = 29000ksi

F_y: Yield strength(ksi) = 50ksi

AISCSteel.Shapes.IShapes.RolledIShapes.MShape — Method

MShape(shape; E=29000ksi, F_y=50ksi, C_b=1)

Constructor for MShape.

AISCSteel.Shapes.IShapes.RolledIShapes.SShape — Type

struct SShape <: AbstractRolledIShapes

SShape in the AISC steel database.

Fields

shape: name of the SShape

weight: weight of section (plf)

area: area of wshape (inch2)

d: depth of wshape (inch)

b_f: width of flange (inch)

t_w: thickness of web (inch)

t_f: thickness of flange (inch)

k: Distance from outer face of flange to web toe of fillet used for design (inch)

k_1: Distance from web center line to flange toe of fillet used for detailing (inch)

h: clear distance between flanges less the fillets (inch)

I_x: Moment of inertia about the x-axis (inch4)

Z_x: Plastic section modulus about the x-axis (inch3)

S_x: Elastic section modulus about the x-axis (inch3)

r_x: Radius of gyration about the x-axis (inch)

I_y: Moment of inertia about the y-axis (inch4)

Z_y: Plastic section modulus about the y-axis (inch3)

S_y: Elastic section modulus about the y-axis (inch3)

r_y: Radius of gyration about the y-axis (inch)

J: Torsional constant (inch4)

C_w: Warping constant (inch6)

W_no: Normalized warping function, as used in Design Guide 9 (inch2)

S_w1: Warping statical moment at point 1 on cross section, as used in AISC Design Guide 9 (inch4)

Q_f: Statical moment for a point in the flange directly above the vertical edge of the web, as used in AISC Design Guide 9 (inch3)

Q_w: Statical moment for a point at mid-depth of the cross section, as used in AISC Design Guide 9 (inch3)

r_ts: Effective radius of gyration (inch)

h_0: Distance between the flange centroids (inch)

PA: Shape perimeter minus one flange surface (or short leg surface for a single angle), as used in Design Guide 19 (inch)

PB: Shape perimeter, as used in AISC Design Guide 19 (inch)

PC: Box perimeter minus one flange surface, as used in Design Guide 19 (inch)

PD: Box perimeter, as used in AISC Design Guide 19 (inch)

T: Distance between web toes od fillets at the top and bottom of web (inch)

WG_i: The workable gage for the inner fastener holes in the flange that provides for entering and tightening clearances and edge distance and spacing requirements. The actual size, combination, and orientation of fastener components should be compared with the geometry of the cross section to ensure compatibility. See AISC Manual Part 1 for additional information (inch)

WG_0: The bolt spacing between inner and outer fastener holes when the workable gage is compatible with four holes across the flange. See AISC Manual Part 1 for additional information (inch)

E: Elastic section modulus (ksi) = 29000ksi

F_y: Yield strength(ksi) = 50ksi

AISCSteel.Shapes.IShapes.RolledIShapes.SShape — Method

SShape(shape; E=29000ksi, F_y=50ksi, C_b=1)

Constructor for SShape.

AISCSteel.Shapes.IShapes.RolledIShapes.WShape — Type

struct WShape <: AbstractRolledIShapes

WShape in the AISC steel database.

Fields

shape: name of the WShape

weight: weight of section (plf)

area: area of wshape (inch2)

d: depth of wshape (inch)

b_f: width of flange (inch)

t_w: thickness of web (inch)

t_f: thickness of flange (inch)

k: Distance from outer face of flange to web toe of fillet used for design (inch)

k_1: Distance from web center line to flange toe of fillet used for detailing (inch)

h: clear distance between flanges less the fillets (inch)

I_x: Moment of inertia about the x-axis (inch4)

Z_x: Plastic section modulus about the x-axis (inch3)

S_x: Elastic section modulus about the x-axis (inch3)

r_x: Radius of gyration about the x-axis (inch)

I_y: Moment of inertia about the y-axis (inch4)

Z_y: Plastic section modulus about the y-axis (inch3)

S_y: Elastic section modulus about the y-axis (inch3)

r_y: Radius of gyration about the y-axis (inch)

J: Torsional constant (inch4)

C_w: Warping constant (inch6)

W_no: Normalized warping function, as used in Design Guide 9 (inch2)

S_w1: Warping statical moment at point 1 on cross section, as used in AISC Design Guide 9 (inch4)

Q_f: Statical moment for a point in the flange directly above the vertical edge of the web, as used in AISC Design Guide 9 (inch3)

Q_w: Statical moment for a point at mid-depth of the cross section, as used in AISC Design Guide 9 (inch3)

r_ts: Effective radius of gyration (inch)

h_0: Distance between the flange centroids (inch)

PA: Shape perimeter minus one flange surface (or short leg surface for a single angle), as used in Design Guide 19 (inch)

PB: Shape perimeter, as used in AISC Design Guide 19 (inch)

PC: Box perimeter minus one flange surface, as used in Design Guide 19 (inch)

PD: Box perimeter, as used in AISC Design Guide 19 (inch)

T: Distance between web toes od fillets at the top and bottom of web (inch)

WG_i: The workable gage for the inner fastener holes in the flange that provides for entering and tightening clearances and edge distance and spacing requirements. The actual size, combination, and orientation of fastener components should be compared with the geometry of the cross section to ensure compatibility. See AISC Manual Part 1 for additional information (inch)

WG_0: The bolt spacing between inner and outer fastener holes when the workable gage is compatible with four holes across the flange. See AISC Manual Part 1 for additional information (inch)

E: Elastic section modulus (ksi) = 29000ksi

F_y: Yield strength(ksi) = 50ksi

AISCSteel.Shapes.IShapes.RolledIShapes.WShape — Method

WShape(shape; E=29000ksi, F_y=50ksi, C_b=1)

Constructor for WShape.

Compression API for IShapes

AISCSteel.Shapes.IShapes.RolledIShapes.Compression — Module

module Compression

This module includes useful functions to calculate compression capacity of rolled i-shape sections (WShape, MShape, SShape, HPShape).

Functions

classify_flange - classify flange for slenderness

classify_web - classify web for slenderness

calc_Pn - Compressive capacity of the shape

AISCSteel.Shapes.IShapes.RolledIShapes.Compression.calc_Pn — Method

calc_Pn(shape::T, L_cx, L_cy) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes
calc_Pn(shape::T, L_cx, L_cy, λ_f, λ_rf, λ_fclass, λ_w, λ_rw, λ_wclass) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates Pn of the shape.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

L_cx: effective length of member for buckling about the x-axis (inch)

L_cy: effective length of member for buckling about the y-axis (inch)

λ_f: slenderness ratio of the flange

λ_rf: nonslender slenderness ratio limit of the flange

λ_fclass: nonslender or slender classification for the flange

λ_w: slenderness ratio of the web

λ_rw: nonslender slenderness ratio limit of the web

λ_wclass: nonslender or slender classification for the web

Returns

P_n: nominal compressive strength of the section (kip)

Reference

AISC Section E3, E7

AISCSteel.Shapes.IShapes.RolledIShapes.Compression.classify_flange — Method

classify_flange(shape::T) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function classifies flange for compression for the shape.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

Returns

(λ_f, λ_rf, λ_fclass)

λ_f: slenderness ratio of the flange

λ_rf: nonslender slenderness ratio limit of the flange

λ_fclass: nonslender or slender classification for the flange

AISCSteel.Shapes.IShapes.RolledIShapes.Compression.classify_web — Method

classify_web(shape::T) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function classifies web for compression for the shape.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

Returns

(λ_w, λ_rw, λ_wclass)

λ_w: slenderness ratio of the web

λ_rw: nonslender slenderness ratio limit of the web

λ_wclass: nonslender or slender classification for the web

Flexure API for IShapes

AISCSteel.Shapes.IShapes.RolledIShapes.Flexure — Module

module Flexure

This module includes useful functions to calculate bending capacity of rolled i-shape sections (WShape, MShape, SShape, HPShape).

Functions

classify_flange_major_axis - classify flange for slenderness when bent about the x-axis

classify_flange_minor_axis - classify flange for slenderness when bent about the y-axis

classify_web - classify web for slnderness

calc_Mnx - moment capacity about the x-axis

calc_Mny - moment capacity about the y-axis

Modules

F2 - includes functions specific for F2 sections

F3 - includes functions specific for F3 sections

F4 - includes functions specific for F4 sections

F5 - includes functions specific for F5 sections

F6 - includes functions specific for F6 sections

AISCSteel.Shapes.IShapes.RolledIShapes.Flexure.calc_Mnx — Method

calc_Mnx(shape::T, L_b, C_b=1) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates Mnx of the shape.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_nx: nominal moment of the section (kip-ft)

Reference

AISC Section F2-F5

AISCSteel.Shapes.IShapes.RolledIShapes.Flexure.calc_Mny — Method

calc_Mny(shape::T) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates Mny of the shape.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

Returns

M_ny: nominal moment of the section (kip-ft)

Reference

AISC Section F6

AISCSteel.Shapes.IShapes.RolledIShapes.Flexure.classify_flange_major_axis — Method

classify_flange_major_axis(shape::T) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function classifies flange for flexure for the shape.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

Returns

(λ_f, λ_pf, λ_rf, λ_fclass)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

AISCSteel.Shapes.IShapes.RolledIShapes.Flexure.classify_flange_minor_axis — Method

classify_flange_minor_axis(shape::T) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function classifies flange for flexure for the shape in minor axis bending.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

Returns

(λ_f, λ_pf, λ_rf, λ_fclass)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

AISCSteel.Shapes.IShapes.RolledIShapes.Flexure.classify_section — Method

classify_section(w::T) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function classifies section for flexure for the shape.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

Returns

section_class: F2, F3, F4, or F5

AISCSteel.Shapes.IShapes.RolledIShapes.Flexure.classify_web — Method

classify_web(shape::T) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function classifies web for flexure for the shape.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

Returns

(λ_w, λ_pw, λ_rw, λ_wclass)

λ_w: slenderness ratio of the web

λ_pw: compact slenderness ratio limit of the web

λ_rw: noncompact slenderness ratio limit of the web

λ_wclass: compact noncompact or slender classification for the web

AISCSteel.ChapterFFlexure.F2.calc_Mn — Method

calc_Mn(shape::T, L_b, C_b=1) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates Mn of the shape for AISC Section F2.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_n: nominal moment of the section (kip-ft)

Reference

AISC Section F2

AISCSteel.ChapterFFlexure.F2.calc_variables — Method

calc_variables(shape::T, L_b, C_b=1) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates the following miscellaneous variables that are used to calculate Mn of the shape for AISC Section F2.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

(; M_p, L_p, L_r, F_cr)

M_p: plastic moment of the section (kip-ft)

L_p: the limiting laterally unbraced length for the limit state of yielding (inch)

L_r: the limiting laterally unbraced length for the limit state of inelastic lateral-torsional buckling (inch)

F_cr: critical stress (ksi)

Reference

AISC Section F2

AISCSteel.ChapterFFlexure.F3.calc_Mn — Method

calc_Mn(shape::T, L_b, C_b=1) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates Mn of the shape for AISC Section F3.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_n: nominal moment of the section (kip-ft)

Reference

AISC Section F3

AISCSteel.ChapterFFlexure.F3.calc_variables — Method

calc_variables(shape::T, L_b, C_b=1) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates the following miscellaneous variables that are used to calculate Mn of the shape for AISC Section F3.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

(; M_p, L_p, L_r, F_cr, k_c)

M_p: plastic moment of the section (kip-ft)

L_p: the limiting laterally unbraced length for the limit state of yielding (inch)

L_r: the limiting laterally unbraced length for the limit state of inelastic lateral-torsional buckling (inch)

F_cr: critical stress (ksi)

k_c:

Reference

AISC Section F3

AISCSteel.ChapterFFlexure.F4.calc_Mn — Method

calc_Mn(shape::T, L_b, C_b=1) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates Mn of the shape for AISC Section F4.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_n: nominal moment of the section (kip-ft)

Reference

AISC Section F4

AISCSteel.ChapterFFlexure.F4.calc_variables — Method

calc_variables(shape::T, L_b, λ_w, λ_pw, λ_rw, C_b=1) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates the following miscellaneous variables that are used to calculate Mn of the shape for AISC Section F4.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

(;M_p, M_yc, M_yt, k_c, F_cr, F_L, L_p, L_r, R_pc, R_pt)

M_p: plastic moment of the section (kip-ft)

M_yc: yield moment in the compression flange

M_yt: yield moment in the tension flange

k_c:

F_cr: critical stress (ksi)

F_L: nominal compression flange stress above which the inelastic buckling limit states apply (ksi)

L_p: the limiting laterally unbraced length for the limit state of yielding (inch)

L_r: the limiting laterally unbraced length for the limit state of inelastic lateral-torsional buckling (inch)

R_pc: web plastification factor

R_pt: web plastification factor corresponding to the tension flange yielding limit state

Reference

AISC Section F4

AISCSteel.ChapterFFlexure.F5.calc_Mn — Method

calc_Mn(shape::T, L_b, C_b=1) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates Mn of the shape for AISC Section F5.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_n: nominal moment of the section (kip-ft)

Reference

AISC Section F5

AISCSteel.ChapterFFlexure.F5.calc_variables — Method

calc_variables(shape::T, L_b, λ_f, λ_pf, λ_rf, λ_fclass, C_b=1) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates the following miscellaneous variables that are used to calculate Mn of the shape for AISC Section F5.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

(;M_p, R_pg, F_crLTB, F_crCFLB)

M_p: plastic moment of the section (kip-ft)

R_pg: bending strength reduction factor

F_crLTB: critical stress relating to lateral torsional buckling

F_crCFLB: critical stress relating to compression flange lateral buckling

Reference

AISC Section F5

AISCSteel.ChapterFFlexure.F6.calc_Mn — Method

calc_Mn(shape::T) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates Mn of the shape for AISC Section F6.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

Returns

M_n: nominal moment of the section (kip-ft)

Reference

AISC Section F6

AISCSteel.ChapterFFlexure.F6.calc_variables — Method

calc_variables(shape::T) where T <: AISCSteel.Shapes.IShapes.AbstractRolledIShapes

This function calculates the following miscellaneous variables that are used to calculate Mn of the shape for AISC Section F6.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

Returns

(;M_p, F_cr)

M_p: plastic moment of the section (kip-ft)

F_cr: critical stress (ksi)

Reference

AISC Section F6