LShapes

AISCSteel.Shapes.LShapes.LShape — Type

struct LShape <: AbstractLShapes

LShape in the AISC steel database. shape,weight,area,d,b,t,k,x,y,xp,yp,Ix,Zx,Sx,rx,Iy,Zy,Sy,ry,Iz,rz,Sz,J,Cw,ro,H,tan_a,Iw,zA,zB,zC,wA,wB,wC,SwA,SwB,SwC,SzA,SzB,SzC,PA,PA2,PB

Fields

shape: name of the WShape

weight: weight of section (plf)

A_g: area of wshape (inch^2)

d: width of shorter leg (inch)

b: width of the longer leg (inch)

t: Thickness of angle leg (inch)

k: Distance from outer face of flange to web toe of fillet used for design (inch)

x̄: Horizontal distance from designated edge of member, as defined in the AISC Steel Construction Manual Part 1, to center of gravity of member (inch)

ȳ: Vertical distance from designated edge of member, as defined in the AISC Steel Construction Manual Part 1, to center of gravity of member (inch)

x_p: Horizontal distance from designated edge of member, as defined in the AISC Steel Construction Manual Part 1, to plastic neutral axis of member (inch)

y_p: Vertical distance from designated edge of member, as defined in the AISC Steel Construction Manual Part 1, to plastic neutral axis of member (inch)

I_x: Moment of inertia about the x-axis (inch^4)

Z_x: Plastic section modulus about the x-axis (inch^3)

S_x: Elastic section modulus about the x-axis (inch^3)

r_x: Radius of gyration about the x-axis (inch)

I_y: Moment of inertia about the y-axis (inch^4)

Z_y: Plastic section modulus about the y-axis (inch^3)

S_y: Elastic section modulus about the y-axis (inch^3)

r_y: Radius of gyration about the y-axis (inch)

r_z: Radius of gyration about the z-axis (inch)

J: Torsional constant (inch^4)

C_w: Warping constant (inch^6)

r_0: Polar radius of gyration about the shear center, in.

S_wA: Elastic section modulus about the w-axis at Point A on the cross section

S_wC: Elastic section modulus about the w-axis at Point C on the cross section

S_zA: Elastic section modulus about the z-axis at Point A on the cross section

S_zB: Elastic section modulus about the z-axis at Point B on the cross section

S_zC: Elastic section modulus about the z-axis at Point C on the cross section

β_w: See F11 commentary for values

E: Elastic section modulus (ksi) = 29000ksi

F_y: Yield strength(ksi) = 50ksi

AISCSteel.Shapes.LShapes.LShape — Method

LShape(shape; E=29000ksi, F_y=50ksi, C_b=1)

Constructor for LShape.

Compression API for LShapes

AISCSteel.Shapes.LShapes.Compression — Module

module Compression

This module includes useful functions to calculate compression capacity of angle sections (LShape).

Functions

classify_leg - classify leg for slenderness

classify_web - classify web for slenderness

calc_Pn - Compressive capacity of the shape

AISCSteel.Shapes.LShapes.Compression.calc_Pn — Method

# assumes connection_type = :type_a
calc_Pn(shape::T, leg_connected, L) where T <: AISCSteel.Shapes.LShapes.AbstractLShapes

# must specify connection_type
calc_Pn(shape::T, connection_type, leg_connected, L) where T <: AISCSteel.Shapes.LShapes.AbstractLShapes
calc_Pn(shape::T, connection_type, leg_connected, L, λ, λ_r, λ_class) where T <: AISCSteel.Shapes.LShapes.AbstractLShapes

This function calculates Pn of the shape.

Arguments

shape: rolled i-shape section (WTShape)

connection_type: the type of connection specified in E5 (type_a or type_b)

leg_connected: the leg that is connected (:short or :long)

L: length of member between work points (inch)

λ: slenderness ratio of the long leg

λ_r: nonslender slenderness ratio limit of the long leg

λ_class: nonslender or slender classification for the long leg

Returns

P_n: nominal compressive strength of the section (kip)

Assumptions

Members are loaded at the ends in compression though the same one leg

Members are attached by welding or by connections with a minimum of two bolts

There are no intermediate transverse loads

Lc/r as determined in this section does not exceed 200

For unequal leg angles, the ratio of long leg width to short leg width is less than 1.7

Reference

AISC Section E3, E5, E7

AISCSteel.Shapes.LShapes.Compression.classify_leg — Method

classify_leg(shape::T) where T <: AISCSteel.Shapes.LShapes.AbstractLShapes

This function classifies flange for compression for the shape.

Arguments

shape: rolled i-shape section (WTShape)

Returns

(λ_f, λ_rf, λ_fclass)

λ_f: slenderness ratio of the flange

λ_rf: nonslender slenderness ratio limit of the flange

λ_fclass: nonslender or slender classification for the flange

Flexure API for LShapes

AISCSteel.Shapes.LShapes.Flexure — Module

module Flexure

This module includes useful functions to calculate bending capacity of LShapes.

Functions

classify_leg

calc_positive_Mnw - moment capacity about w-axis when compression is in the short leg.

calc_negative_Mnw - moment capacity about w-axis when compression is in the long leg.

calc_positive_Mnz - moment capacity about z-axis when compression is in the toe of the legs.

calc_negative_Mnz - moment capacity about z-axis when tension is in the toe of the legs.

calc_positive_Mnx - moment capacity about x-axis when compression is in the toe of the leg.

calc_negative_Mnx - moment capacity about x-axis when tension is in the toe of the leg.

calc_positive_Mny - moment capacity about y-axis when compression is in the toe of the leg.

calc_negative_Mny - moment capacity about y-axis when tension is in the toe of the leg.

Modules

F10 - module where functions are exported from. No need to step into this module

AISCSteel.Shapes.LShapes.Flexure.classify_leg — Method

classify_leg(shape::T) where T <: AISCSteel.Shapes.LShapes.AbstractLShapes

This function classifies leg for flexure for the shape.

Arguments

shape: rolled L-Shape section (LShape)

Returns

(λ_f, λ_p, λ_r, λ_class)

λ: slenderness ratio of the leg

λ_p: compact slenderness ratio limit of the leg

λ_rf: noncompact slenderness ratio limit of the leg

λ_fclass: compact noncompact or slender classification for the leg

AISCSteel.Shapes.LShapes.Flexure.F10.GeometricAxisBending — Module

module GeometricAxisBending

LShapes bent about their geometric axis (x-axis, y-axis).

There are two sections:

Positive Bending - when compression is in the toe of the leg

Negative Bending - when tension is in the toe of the leg

AISCSteel.Shapes.LShapes.Flexure.F10.GeometricAxisBending.PositiveBending — Module

module PositiveBending

LShapes bent about their geometric axis (x-axis, y-axis) when compression is in the toe of the leg.

AISCSteel.Shapes.LShapes.Flexure.F10.GeometricAxisBending.PositiveBending.calc_positive_Mnx — Method

calc_positive_Mnx(lshape, λ_class, L_b, restraint_type, C_b)
calc_positive_Mnx(lshape, L_b, restraint_type, C_b)

Calculates positive moment (when toe of leg is in compression) about geometric axis for an LShape.

Arguments

lshape: LShape object

λ_class: slenderness classification of angle leg

L_b: unbraced length (inch)

restraint_type: type of restraint on leg (:fully_restrained, :unrestrained, or :at_max_moment_only)

C_b: lateral torsional buckling modification factor

AISCSteel.Shapes.LShapes.Flexure.F10.GeometricAxisBending.PositiveBending.calc_positive_Mny — Method

calc_positive_Mny(lshape, λ_class, L_b, restraint_type, C_b)
calc_positive_Mny(lshape, L_b, restraint_type, C_b)

Calculates positive moment (when toe of leg is in compression) about geometric axis for an LShape.

Arguments

lshape: LShape object

λ_class: slenderness classification of angle leg

L_b: unbraced length (inch)

restraint_type: type of restraint on leg (:fully_restrained, :unrestrained, or :at_max_moment_only)

C_b: lateral torsional buckling modification factor

AISCSteel.Shapes.LShapes.Flexure.F10.GeometricAxisBending.NegativeBending — Module

module NegativeBending

LShapes bent about their geometric axis (x-axis, y-axis) when tension is in the toe of the leg.

AISCSteel.Shapes.LShapes.Flexure.F10.GeometricAxisBending.NegativeBending.calc_negative_Mnx — Method

calc_negative_Mnx(lshape, L_b, restraint_type, C_b)

Calculates negative moment (when toe of leg is in tension) about geometric axis for an LShape.

Arguments

lshape: LShape object

L_b: unbraced length (inch)

restraint_type: type of restraint on leg (:fully_restrained, :unrestrained, or :at_max_moment_only)

C_b: lateral torsional buckling modification factor

AISCSteel.Shapes.LShapes.Flexure.F10.GeometricAxisBending.NegativeBending.calc_negative_Mny — Method

calc_negative_Mny(lshape, L_b, restraint_type, C_b)

Calculates negative moment (when toe of leg is in tension) about geometric axis for an LShape.

Arguments

lshape: LShape object

L_b: unbraced length (inch)

restraint_type: type of restraint on leg (:fully_restrained, :unrestrained, or :at_max_moment_only)

C_b: lateral torsional buckling modification factor

AISCSteel.Shapes.LShapes.Flexure.F10.PrincipalAxisBending.MajorAxis — Module

module MajorAxis

LShapes bent about their principal major axis (w-axis).

There are two sections:

Positive Bending - when compression is in the short leg

Negative Bending - when compression is in the long leg

AISCSteel.Shapes.LShapes.Flexure.F10.PrincipalAxisBending.MajorAxis.PositiveBending — Module

module PositiveBending

LShapes bent about their principal major axis (w-axis) when compression is in the short leg.

AISCSteel.Shapes.LShapes.Flexure.F10.PrincipalAxisBending.MajorAxis.NegativeBending — Module

module NegativeBending

LShapes bent about their principal major axis (w-axis) when compression is in the long leg.

AISCSteel.Shapes.LShapes.Flexure.F10.PrincipalAxisBending.MinorAxis — Module

module MinorAxis

LShapes bent about their principal minor axis (z-axis).

There are two sections:

Positive Bending - when tension is in the toe of the legs

Negative Bending - when compression is in the toe of the legs

AISCSteel.Shapes.LShapes.Flexure.F10.PrincipalAxisBending.MinorAxis.PositiveBending — Module

module PositiveBending

LShapes bent about their principal minor axis (z-axis) when tension is in the toe of the legs.

AISCSteel.Shapes.LShapes.Flexure.F10.PrincipalAxisBending.MinorAxis.NegativeBending — Module

module NegativeBending

LShapes bent about their principal minor axis (z-axis) when compression is in the toe of the legs.