Round HSS Shapes

AISCSteel.Shapes.RoundHSS_Shapes.RoundHSS_Shape — Type

struct RoundHSS_Shape <: AbstractHSS_Shapes

RoundHSS_Shape in the AISC steel database.

Fields

shape: name of the HSS

weight: weight of section (plf)

A_g: area of shape (inch2)

OD: outside diameter of round HSS or pipe (inch)

t_nom: nominal thickness of HSS and pipe wall (inch)

t_des: design thickness of HSS and pipe wall (inch)

I_x: Moment of inertia about the x-axis (inch4)

Z_x: Plastic section modulus about the x-axis (inch3)

S_x: Elastic section modulus about the x-axis (inch3)

r_x: Radius of gyration about the x-axis (inch)

I_y: Moment of inertia about the y-axis (inch4)

Z_y: Plastic section modulus about the y-axis (inch3)

S_y: Elastic section modulus about the y-axis (inch3)

r_y: Radius of gyration about the y-axis (inch)

J: Torsional constant (inch4)

C: HSS torsional constant (inch3)

E: Elastic section modulus (ksi) = 29000ksi

F_y: Yield strength(ksi) = 50ksi

AISCSteel.Shapes.RoundHSS_Shapes.RoundHSS_Shape — Method

RoundHSS_Shape(shape; E=29000ksi, F_y=50ksi, C_b=1)

Constructor for RoundHSS_Shape.

Compression API for Round HSS Shapes

AISCSteel.Shapes.RoundHSS_Shapes.Compression — Module

module Compression

This module includes useful functions to calculate compression capacity of round hss-shape sections (RoundHSS_Shape).

Functions

classify_wall - classify wall for slenderness

calc_Pn - Compressive capacity of the shape

AISCSteel.Shapes.RoundHSS_Shapes.Compression.calc_Pn — Method

calc_Pn(shape::T, L_cx, L_cy) where T <: AISCSteel.Shapes.RoundHSS_Shapes.AbstractRoundHSS_Shapes
calc_Pn(shape::T, L_cx, L_cy, λ_b, λ_rb, λ_bclass, λ_h, λ_rh, λ_hclass) where T <: AISCSteel.Shapes.RoundHSS_Shapes.AbstractRoundHSS_Shapes

This function calculates Pn of the shape.

Arguments

shape: rolled hss-shape section (RoundHSS_Shape)

L_cx: effective length of member for buckling about the x-axis (inch)

L_cy: effective length of member for buckling about the y-axis (inch)

λ: slenderness ratio of the wall

λ_r: nonslender slenderness ratio limit of the wall

λ_class: nonslender or slender classification for the wall

Returns

P_n: nominal compressive strength of the section (kip)

Reference

AISC Section E3, E7

AISCSteel.Shapes.RoundHSS_Shapes.Compression.classify_wall — Method

classify_wall(shape::T) where T <: AISCSteel.Shapes.RoundHSS_Shapes.AbstractRoundHSS_Shapes

This function classifies wall for compression for the shape.

Arguments

shape: rolled hss-shape section (RoundHSS_Shape)

Returns

(λ_f, λ_rf, λ_fclass)

λ_f: slenderness ratio of the wall

λ_rf: nonslender slenderness ratio limit of the wall

λ_fclass: nonslender or slender classification for the wall

Flexure API for Round HSS Shapes

AISCSteel.Shapes.RoundHSS_Shapes.Flexure — Module

module Flexure

This module includes useful functions to calculate bending capacity of rolled HSS sections (HSS_Shape).

Functions

classify_round_hss - classify round hss for slenderness

calc_Mn - moment capacity

AISCSteel.Shapes.RoundHSS_Shapes.Flexure.calc_Mn — Method

calc_Mn(shape::T) where T <: AISCSteel.Shapes.RoundHSS_Shapes.AbstractRoundHSS_Shapes
calc_Mn(shape::T, λ_class) where T <: AISCSteel.Shapes.RoundHSS_Shapes.AbstractRoundHSS_Shapes

This function calculates Mnx of the shape.

Arguments

shape: HSS section (HSS_Shape)

λ_class: compact noncompact or slender classification for the round hss

Returns

M_n: nominal moment of the section (kip-ft)

Reference

AISC Section F8

AISCSteel.Shapes.RoundHSS_Shapes.Flexure.classify_round_hss — Method

classify_round_hss(shape::T) where T <: AISCSteel.Shapes.HSS_Shapes.AbstractHSS_Shapes

This function classifies flange for flexure for the shape.

Arguments

shape: rolled HSS sections (HSS_Shape)

Returns

(λ_f, λ_pf, λ_rf, λ_fclass)

λ_f: slenderness ratio of the round hss

λ_pf: compact slenderness ratio limit of the round hss

λ_rf: noncompact slenderness ratio limit of the round hss

λ_fclass: compact noncompact or slender classification for the round hss