WTShapes

AISCSteel.Shapes.WTShapes.MTShape — Type

struct MTShape <: AbstractWTShapes

MTShape in the AISC steel database.

Fields

shape: name of the WShape

weight: weight of section (plf)

A_g: area of wshape (inch2)

d: depth of wshape (inch)

b_f: width of flange (inch)

t_w: thickness of web (inch)

t_f: thickness of flange (inch)

k: Distance from outer face of flange to web toe of fillet used for design (inch)

ȳ:

y_p:

I_x: Moment of inertia about the x-axis (inch4)

Z_x: Plastic section modulus about the x-axis (inch3)

S_x: Elastic section modulus about the x-axis (inch3)

r_x: Radius of gyration about the x-axis (inch)

I_y: Moment of inertia about the y-axis (inch4)

Z_y: Plastic section modulus about the y-axis (inch3)

S_y: Elastic section modulus about the y-axis (inch3)

r_y: Radius of gyration about the y-axis (inch)

J: Torsional constant (inch4)

C_w: Warping constant (inch6)

r_0:

H:

G: Shear modulus of elasticity of steel = 11200ksi

E: Elastic section modulus (ksi) = 29000ksi

F_y: Yield strength(ksi) = 50ksi

AISCSteel.Shapes.WTShapes.MTShape — Method

MTShape(shape; E=29000ksi, F_y=50ksi, C_b=1)

Constructor for MTShape.

AISCSteel.Shapes.WTShapes.STShape — Type

struct STShape <: AbstractWTShapes

STShape in the AISC steel database.

Fields

shape: name of the WShape

weight: weight of section (plf)

A_g: area of wshape (inch2)

d: depth of wshape (inch)

b_f: width of flange (inch)

t_w: thickness of web (inch)

t_f: thickness of flange (inch)

k: Distance from outer face of flange to web toe of fillet used for design (inch)

ȳ:

y_p:

I_x: Moment of inertia about the x-axis (inch4)

Z_x: Plastic section modulus about the x-axis (inch3)

S_x: Elastic section modulus about the x-axis (inch3)

r_x: Radius of gyration about the x-axis (inch)

I_y: Moment of inertia about the y-axis (inch4)

Z_y: Plastic section modulus about the y-axis (inch3)

S_y: Elastic section modulus about the y-axis (inch3)

r_y: Radius of gyration about the y-axis (inch)

J: Torsional constant (inch4)

C_w: Warping constant (inch6)

r_0:

H:

G: Shear modulus of elasticity of steel = 11200ksi

E: Elastic section modulus (ksi) = 29000ksi

F_y: Yield strength(ksi) = 50ksi

AISCSteel.Shapes.WTShapes.STShape — Method

STShape(shape; E=29000ksi, F_y=50ksi, C_b=1)

Constructor for STShape.

AISCSteel.Shapes.WTShapes.WTShape — Type

struct WTShape <: AbstractWTShapes

WTShape in the AISC steel database.

Fields

shape: name of the WShape

weight: weight of section (plf)

A_g: area of wshape (inch2)

d: depth of wshape (inch)

b_f: width of flange (inch)

t_w: thickness of web (inch)

t_f: thickness of flange (inch)

k: Distance from outer face of flange to web toe of fillet used for design (inch)

ȳ:

y_p:

I_x: Moment of inertia about the x-axis (inch4)

Z_x: Plastic section modulus about the x-axis (inch3)

S_x: Elastic section modulus about the x-axis (inch3)

r_x: Radius of gyration about the x-axis (inch)

I_y: Moment of inertia about the y-axis (inch4)

Z_y: Plastic section modulus about the y-axis (inch3)

S_y: Elastic section modulus about the y-axis (inch3)

r_y: Radius of gyration about the y-axis (inch)

J: Torsional constant (inch4)

C_w: Warping constant (inch6)

r_0:

H:

PA: Shape perimeter minus one flange surface (or short leg surface for a single angle), as used in Design Guide 19 (inch)

PB: Shape perimeter, as used in AISC Design Guide 19 (inch)

PC: Box perimeter minus one flange surface, as used in Design Guide 19 (inch)

PD: Box perimeter, as used in AISC Design Guide 19 (inch)

WG_i: The workable gage for the inner fastener holes in the flange that provides for entering and tightening clearances and edge distance and spacing requirements. The actual size, combination, and orientation of fastener components should be compared with the geometry of the cross section to ensure compatibility. See AISC Manual Part 1 for additional information (inch)

WG_0: The bolt spacing between inner and outer fastener holes when the workable gage is compatible with four holes across the flange. See AISC Manual Part 1 for additional information (inch)

G: Shear modulus of elasticity of steel = 11200ksi

E: Elastic section modulus (ksi) = 29000ksi

F_y: Yield strength(ksi) = 50ksi

AISCSteel.Shapes.WTShapes.WTShape — Method

WTShape(shape; E=29000ksi, F_y=50ksi, C_b=1)

Constructor for WTShape.

Compression API for WTShapes

AISCSteel.Shapes.WTShapes.Compression — Module

module Compression

This module includes useful functions to calculate compression capacity of rolled i-shape sections (WShape, MShape, SShape, HPShape).

Functions

classify_flange - classify flange for slenderness

classify_web - classify web for slenderness

calc_Pn - Compressive capacity of the shape

AISCSteel.Shapes.WTShapes.Compression.calc_Fe — Method

calc_Fe(F_ey, F_ez, H)

This function calculates F_e of the shape.

Arguments

F_ey: elastic buckling stress with respect to the y-axis (ksi)

F_ez: elastic buckling stress with respect to the z-axis (ksi)

H: flexural constant

Returns

F_e: elastic buckling stress (ksi)

Reference

AISC Section E4 (E4-3)

AISCSteel.Shapes.WTShapes.Compression.calc_Pn — Method

calc_Pn(shape::T, L_cx, L_cy, L_cz) where T <: AISCSteel.Shapes.WTShapes.AbstractWTShapes
calc_Pn(shape::T, L_cx, L_cy, L_cz, λ_f, λ_rf, λ_fclass, λ_w, λ_rw, λ_wclass) where T <: AISCSteel.Shapes.WTShapes.AbstractWTShapes

This function calculates Pn of the shape.

Arguments

shape: rolled i-shape section (WTShape)

L_cx: effective length of member for buckling about the x-axis (inch)

L_cy: effective length of member for buckling about the y-axis (inch)

λ_f: slenderness ratio of the flange

λ_rf: nonslender slenderness ratio limit of the flange

λ_fclass: nonslender or slender classification for the flange

λ_w: slenderness ratio of the web

λ_rw: nonslender slenderness ratio limit of the web

λ_wclass: nonslender or slender classification for the web

Returns

P_n: nominal compressive strength of the section (kip)

Reference

AISC Section E3, E4, E7

AISCSteel.Shapes.WTShapes.Compression.calc_r̄0 — Method

calc_r̄0(x_0, y_0, I_x, I_y, A_g)

Calculates the polar radius of gyration about the shear center.

Description of applicable member: member without slender elements.

Arguments

x_0: x-coordinate of the shear center with respect to the centroid (inch)

y_0: y-coordinate of the shear center with respect to the centroid (inch)

I_x: Moment of inertia about the x-axis (inch^4)

I_y: Moment of inertia about the y-axis (inch^4)

A_g: gross area of member (inch^2)

Returns

r̄_0: polar radius of gyration about the shear center (inch)

Reference

AISC Section E4 (E4-8)

AISCSteel.Shapes.WTShapes.Compression.classify_flange — Method

classify_flange(shape::T) where T <: AISCSteel.Shapes.WTShapes.AbstractWTShapes

This function classifies flange for compression for the shape.

Arguments

shape: rolled i-shape section (WTShape)

Returns

(λ_f, λ_rf, λ_fclass)

λ_f: slenderness ratio of the flange

λ_rf: nonslender slenderness ratio limit of the flange

λ_fclass: nonslender or slender classification for the flange

AISCSteel.Shapes.WTShapes.Compression.classify_web — Method

classify_web(shape::T) where T <: AISCSteel.Shapes.WTShapes.AbstractWTShapes

This function classifies web for compression for the shape.

Arguments

shape: rolled i-shape section (WShape, MShape, SShape, HPShape)

Returns

(λ_w, λ_rw, λ_wclass)

λ_w: slenderness ratio of the web

λ_rw: nonslender slenderness ratio limit of the web

λ_wclass: nonslender or slender classification for the web

Flexure API for WTShapes

AISCSteel.Shapes.WTShapes.Flexure.calc_negative_Mnx — Method

calc_negative_Mnx(shape::T, L_b) where T <: AISCSteel.Shapes.WTShapes.AbstractWTShapes

This function calculates Mnx of the shape.

Arguments

shape: rolled WT-Shape section (WTShape)

L_b: unbraced length (inch)

Returns

M_nx: nominal moment of the section (kip-ft)

Reference

AISC Section F9

AISCSteel.Shapes.WTShapes.Flexure.calc_positive_Mnx — Method

calc_positive_Mnx(shape::T, L_b, λ_f, λ_pf, λ_rf, λ_fclass) where T <: AISCSteel.Shapes.WTShapes.AbstractWTShapes

This function calculates Mnx of the shape.

Arguments

shape: rolled WT-Shape section (WTShape)

L_b: unbraced length (inch)

Returns

M_nx: nominal moment of the section (kip-ft)

Reference

AISC Section F9

AISCSteel.Shapes.WTShapes.Flexure.classify_flange — Method

classify_flange(shape::T) where T <: AISCSteel.Shapes.WTShapes.AbstractWTShapes

This function classifies flange for flexure for the shape.

Arguments

shape: rolled WT-Shape section (WTShape)

Returns

(λ_f, λ_pf, λ_rf, λ_fclass)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

AISCSteel.ChapterFFlexure.F9.WT.NegativeBending.calc_Mn — Method

calc_Mn(shape::T) where T <: AISCSteel.Shapes.WTShapes.AbstractWTShapes

This function calculates Mn of the shape for AISC Section F6.

Arguments

shape: rolled WT-Shape section (WTShape)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

Returns

M_n: nominal moment of the section (kip-ft)

Reference

AISC Section F9

AISCSteel.ChapterFFlexure.F9.WT.NegativeBending.calc_variables — Method

calc_variables(shape::T) where T <: AISCSteel.Shapes.WTShapes.AbstractWTShapes

This function calculates the following miscellaneous variables that are used to calculate Mn of the shape for AISC Section F6.

Arguments

shape: rolled WT-Shape section (WTShape)

Returns

(;M_p, F_cr)

M_p: plastic moment of the section (kip-ft)

F_cr: critical stress (ksi)

Reference

AISC Section F9

AISCSteel.ChapterFFlexure.F9.WT.PositiveBending.calc_Mn — Method

calc_Mn(shape::T) where T <: AISCSteel.Shapes.WTShapes.AbstractWTShapes

This function calculates Mn of the shape for AISC Section F6.

Arguments

shape: rolled WT-Shape section (WTShape)

L_b: unbraced length (inch)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

Returns

M_n: nominal moment of the section (kip-ft)

Reference

AISC Section F9

AISCSteel.ChapterFFlexure.F9.WT.PositiveBending.calc_variables — Method

calc_variables(shape::T) where T <: AISCSteel.Shapes.WTShapes.AbstractWTShapes

This function calculates the following miscellaneous variables that are used to calculate Mn of the shape for AISC Section F6.

Arguments

shape: rolled WT-Shape section (WTShape)

L_b: lateral unbraced length

Returns

(;M_p, F_cr)

M_p: plastic moment of the section (kip-ft)

F_cr: critical stress (ksi)

Reference

AISC Section F9