AISCSteel

Documentation for AISCSteel.

AISCSteel.classify_section_for_lb_case1

AISCSteel.classify_section_for_lb_case10

AISCSteel.classify_section_for_lb_case13

AISCSteel.classify_section_for_lb_case15

AISCSteel.flexure_capacity_f2

AISCSteel.flexure_capacity_f2_1

AISCSteel.flexure_capacity_f2_2

AISCSteel.flexure_capacity_f3

AISCSteel.flexure_capacity_f3_1

AISCSteel.flexure_capacity_f3_2

AISCSteel.flexure_capacity_f4

AISCSteel.flexure_capacity_f4_1

AISCSteel.flexure_capacity_f4_2

AISCSteel.flexure_capacity_f4_3

AISCSteel.flexure_capacity_f4_4

AISCSteel.flexure_capacity_f5

AISCSteel.flexure_capacity_f5_1

AISCSteel.flexure_capacity_f5_2

AISCSteel.flexure_capacity_f5_3

AISCSteel.flexure_capacity_f5_4

AISCSteel.import_data

AISCSteel.classify_section_for_lb_case1 — Method

classify_section_for_lb_case1(b, t, E, F_y)

Classifies the section of a member subject to axial compression.

Description of applicable member: Flanges of rolled I-shaped sections, plates projecting from rolled I-shaped sections, outstanding legs of pairs of angles connected with continuous contact, flanges of channels, and flanges of tees.

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

CompressionBucklingType enum whose value is either Nonslender or Slender

Reference

AISC Table B4.1a

AISCSteel.classify_section_for_lb_case10 — Method

classify_section_for_lb_case10(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Flanges of rolled I-shaped sections, channels, and tees.

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.classify_section_for_lb_case13 — Method

classify_section_for_lb_case13(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Flanges of all I-shaped sections and channels bent about their minor axis.

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.classify_section_for_lb_case15 — Method

classify_section_for_lb_case15(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Webs of doubly-symmetric I-shaped sections and channels.

Arguments

h:

For rolled sections: the clear distance between flanges less the fillet at each flange.

For built-up sections: clear distance between flanges when welds are used.

t_w: thickness of the web.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.flexure_capacity_f2 — Function

flexure_capacity_f2(E, F_y, Z_x, S_x, r_y, h_0, J, c, r_ts, L_b, C_b=1)

Calculates the moment capacity of the applicable section.

Description of applicable member: Doubly symmetric compact I-shaped members and channels bent about their major axis.

Arguments

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

Z_x: plastic section modulous (inch^4)

S_x: elastic section modulous (inch^3)

r_y: radius of gyration about the y-axis (inch)

h_0: distance between the flange centroids (inch)

J: torsional constant (inch^4)

c:

r_ts:

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_n: moment capacity of the section.

Reference

AISC Section F2

AISCSteel.flexure_capacity_f2_1 — Method

flexure_capacity_f2_1(M_p)

Calculates the moment capacity of the applicable section for yielding.

Description of applicable member: Doubly symmetric compact I-shaped members and channels bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

Returns

M_nFY: moment capacity of the section for yielding.

Reference

AISC Section F2.1

AISCSteel.flexure_capacity_f2_2 — Function

flexure_capacity_f2_2(M_p, F_y, S_x, F_cr, L_b, L_p, L_r, C_b=1)

Calculates the moment capacity of the applicable section for lateral torsional buckling.

Description of applicable member: Doubly symmetric compact I-shaped members and channels bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

F_y: yield strength of steel (ksi)

S_x: elastic section modulous (inch^3)

F_cr: critical stress (ksi)

L_b: unbraced length (inch)

L_p: the limiting laterally unbraced length for the limit state of yielding (inch)

L_r: the limiting laterally unbraced length for the limit state of inelastic lateral-torsional buckling (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_nLTB: moment capacity of the section for lateral torsional buckling.

Reference

AISC Section F2.2

AISCSteel.flexure_capacity_f3 — Function

flexure_capacity_f3(E, F_y, Z_x, S_x, r_y, h_0, J, c, r_ts, L_b, h, t_w, λ_f, λ_pf, λ_rf, λ_fclass, C_b=1)

Calculates the moment capacity of the applicable section.

Description of applicable member: Doubly symmetric I-shaped members with compact webs and noncompact or slender flanges bent about their major axis.

Arguments

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

Z_x: plastic section modulous (inch^4)

S_x: elastic section modulous (inch^3)

r_y: radius of gyration about the y-axis (inch)

h_0: distance between the flange centroids (inch)

J: torsional constant (inch^4)

c:

r_ts:

L_b: unbraced length (inch)

h: clear distance between flanges less the fillets (inch)

t_w: thickness of the web (inch)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_n: moment capacity of the section.

Reference

AISC Section F3

AISCSteel.flexure_capacity_f3_1 — Function

flexure_capacity_f3_1(M_p, F_y, S_x, F_cr, L_b, L_p, L_r, C_b=1)

Calculates the moment capacity of the applicable section for yielding.

Description of applicable member: Doubly symmetric I-shaped members with compact webs and noncompact or slender flanges bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

F_y: yield strength of steel (ksi)

S_x: elastic section modulous (inch^3)

F_cr: critical stress (ksi)

L_b: unbraced length (inch)

L_p: the limiting laterally unbraced length for the limit state of yielding (inch)

L_r: the limiting laterally unbraced length for the limit state of inelastic lateral-torsional buckling (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_nLTB: moment capacity of the section for lateral torsional buckling.

Reference

AISC Section F3.1

AISCSteel.flexure_capacity_f3_2 — Method

flexure_capacity_f3_2(M_p, E, F_y, S_x, k_c, λ_f, λ_pf, λ_rf, λ_fclass)

Calculates the moment capacity of the applicable section for compression flange local buckling.

Description of applicable member: Doubly symmetric I-shaped members with compact webs and noncompact or slender flanges bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

S_x: elastic section modulous (inch^3)

k_c:

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

Returns

M_nCFLB: moment capacity of the section for compression flange local buckling.

Reference

AISC Section F3.2

AISCSteel.flexure_capacity_f4 — Method

flexure_capacity_f4(E, F_y, Z_x, S_x, S_xc, S_xt, b_fc, t_fc, h, h_c, t_w, J, h_0, I_y, I_yc, λ_w, λ_pw, λ_rw, λ_f, λ_pf, λ_rf, λ_fclass, L_b, C_b=1)

Calculates the moment capacity of the applicable section.

Description of applicable member: Other I-shaped members with compact webs or noncompact webs bent about their major axis.

Arguments

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

Z_x: plastic section modulous (inch^4)

S_x: elastic section modulous (inch^3)

S_xc: elastic section modulous referred to compression flange (inch^3)

S_xt: elastic section modulous referred to tension flange (inch^3)

b_fc: width of compression flange (inch)

t_fc: thickness of compression flange (inch)

h: clear distance between flanges less the fillets (inch)

h_c: twice the distance from the centroid to the following: (inch)

the inside face of the compression flange less the fillet or corner radius for rolled shapes

the nearest line of fasteners at the compression flange or inside face of the compression flange when welds are used for built-up sections

t_w: thickness of the web (inch)

J: torsional constant (inch^4)

h_0: distance between the flange centroids (inch)

I_y: second moment of inertia about the y-axis (inch^4)

I_yc: second moment of inertia about the y-axis for the compression flange (inch^4)

λ_w: slenderness ratio of the flange

λ_pw: compact slenderness ratio limit of the web

λ_rw: noncompact slenderness ratio limit of the web

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_n: moment capacity of the section.

Reference

AISC Section F4

AISCSteel.flexure_capacity_f4_1 — Method

flexure_capacity_f4_1(R_pc, M_yc)

Calculates the moment capacity of the applicable section for compression flange yielding.

Description of applicable member: Other I-shaped members with compact webs or noncompact webs bent about their major axis.

Arguments

R_pc: web plastification factor

M_yc: yield moment in the compression flange

Returns

M_nCFY: moment capacity of the section for compression flange yielding.

Reference

AISC Section F4.1

AISCSteel.flexure_capacity_f4_2 — Function

flexure_capacity_f4_2(M_p, R_pc, M_yc, F_L, S_xc, F_cr, L_b, L_p, L_r, C_b=1)

Calculates the moment capacity of the applicable section for lateral torsional buckling.

Description of applicable member: Other I-shaped members with compact webs or noncompact webs bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

R_pc: web plastification factor

M_yc: yield moment in the compression flange

F_L: nominal compression flange stress above which the inelastic buckling limit states apply (ksi)

S_xc: elastic section modulous referred to compression flange (inch^3)

F_cr: critical stress (ksi)

L_b: unbraced length (inch)

L_p: the limiting laterally unbraced length for the limit state of yielding (inch)

L_r: the limiting laterally unbraced length for the limit state of inelastic lateral-torsional buckling (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_nLTB: moment capacity of the section for lateral torsional buckling.

Reference

AISC Section F4.2

AISCSteel.flexure_capacity_f4_3 — Method

flexure_capacity_f4_3(M_p, R_pc, M_yc, F_L, S_xc, E, k_c, λ_f, λ_pf, λ_rf, λ_fclass)

Calculates the moment capacity of the applicable section for compression flange local buckling.

Description of applicable member: Other I-shaped members with compact webs or noncompact webs bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

R_pc: web plastification factor

M_yc: yield moment in the compression flange

F_L: nominal compression flange stress above which the inelastic buckling limit states apply (ksi)

S_xc: elastic section modulous referred to compression flange (inch^3)

E: modulous of elasticity (ksi)

k_c:

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

Returns

M_nCFLB: moment capacity of the section for compression flange local buckling.

Reference

AISC Section F4.3

AISCSteel.flexure_capacity_f4_4 — Method

flexure_capacity_f4_4(M_p, R_pt, M_yt, S_xc, S_xt)

Calculates the moment capacity of the applicable section for tension flange yielding.

Description of applicable member: Other I-shaped members with compact webs or noncompact webs bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

R_pt: web plastification factor corresponding to the tension flange yielding limit state

M_yt: yield moment in the tension flange

S_xc: elastic section modulous referred to compression flange (inch^3)

S_xt: elastic section modulous referred to tension flange (inch^3)

Returns

M_nTFY: moment capacity of the section for tension flange yielding.

Reference

AISC Section F4.4

AISCSteel.flexure_capacity_f5 — Function

flexure_capacity_f5(E, F_y, S_xc, S_xt, b_fc, t_fc, h, h_c, t_w, λ_f, λ_pf, λ_rf, λ_fclass, L_b, C_b=1)

Calculates the moment capacity of the applicable section.

Description of applicable member: I-shaped members with slender webs bent about their major axis.

Arguments

E: modulous of elasticity (ksi)

F_y: yield strength of steel (ksi)

S_x: elastic section modulous (inch^3)

S_xc: elastic section modulous referred to compression flange (inch^3)

S_xt: elastic section modulous referred to tension flange (inch^3)

b_fc: width of compression flange (inch)

t_fc: thickness of compression flange (inch)

h: clear distance between flanges less the fillets (inch)

h_c: twice the distance from the centroid to the following: (inch)

the inside face of the compression flange less the fillet or corner radius for rolled shapes

the nearest line of fasteners at the compression flange or inside face of the compression flange when welds are used for built-up sections

t_w: thickness of the web (inch)

λ_f: slenderness ratio of the flange

λ_pf: compact slenderness ratio limit of the flange

λ_rf: noncompact slenderness ratio limit of the flange

λ_fclass: compact noncompact or slender classification for the flange

L_b: unbraced length (inch)

C_b: lateral torsional buckling modification factor (default = 1)

Returns

M_n: moment capacity of the section.

Reference

AISC Section F5

AISCSteel.flexure_capacity_f5_1 — Method

flexure_capacity_f5_1(M_p)

Calculates the moment capacity of the applicable section for compression flange yielding.

Description of applicable member: I-shaped members with slender webs bent about their major axis.

Arguments

M_p: plastic moment of the section (kip-ft)

Returns

M_nCFY: moment capacity of the section for compression flange yielding.

Reference

AISC Section F5.1

AISCSteel.flexure_capacity_f5_2 — Method

flexure_capacity_f5_2(R_pg, F_crLTB, S_xc)

Calculates the moment capacity of the applicable section for lateral torsional buckling.

Description of applicable member: I-shaped members with slender webs bent about their major axis.

Arguments

R_pg: bending strength reduction factor

F_crLTB: critical stress relating to lateral torsional buckling

S_xc: elastic section modulous referred to compression flange (inch^3)

Returns

M_nLTB: moment capacity of the section for lateral torsional buckling.

Reference

AISC Section F5.2

AISCSteel.flexure_capacity_f5_3 — Method

flexure_capacity_f5_3(R_pg, F_crCFLB, S_xc)

Calculates the moment capacity of the applicable section for compression flange lateral buckling.

Description of applicable member: I-shaped members with slender webs bent about their major axis.

Arguments

R_pg: bending strength reduction factor

F_crCFLB: critical stress relating to compression flange lateral buckling

S_xc: elastic section modulous referred to compression flange (inch^3)

Returns

M_nCFLB: moment capacity of the section for compression flange lateral buckling.

Reference

AISC Section F5.3

AISCSteel.flexure_capacity_f5_4 — Method

flexure_capacity_f5_4(F_y, S_xt)

Calculates the moment capacity of the applicable section for tension flange yielding.

Description of applicable member: I-shaped members with slender webs bent about their major axis.

Arguments

F_y: yield strength of steel (ksi)

S_xt: elastic section modulous referred to tension flange (inch^3)

Returns

M_nTFY: moment capacity of the section for tension flange yielding.

Reference

AISC Section F5.4

AISCSteel.import_data — Method

import_data(lookup_value, lookup_col_name::Symbol, data_location::String)

Searches the specified data location and filters for the first row of data found where the value in the column specified is equal to the lookup value.

Parameters

lookup_value - The value to search for

lookup_col_name::Symbol - the column name to search in

csv_file_path::String - The name of csv file in the data directory