Chapter B - Design Requirements

B4. Member Properties

Classification of Sections for Local Buckling

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case1 — Method

case1(b, t, E, F_y)

Classifies the section of a member subject to axial compression.

Description of applicable member: Flanges of rolled I-shaped sections, plates projecting from rolled I-shaped sections, outstanding legs of pairs of angles connected with continuous contact, flanges of channels, and flanges of tees.

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

CompressionBucklingType enum whose value is either Nonslender or Slender

Reference

AISC Table B4.1a

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case10 — Method

case10(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Flanges of rolled I-shaped sections, channels, and tees.

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case11 — Method

case11(b, t, E, F_L, k_c)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Flanges of built-up I-shaped sections (singly or doubly symmetric).

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_L: nominal compression flange stress above which the inelastic buckling limit states apply (ksi)

k_c:

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case13 — Method

case13(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Flanges of all I-shaped sections and channels bent about their minor axis.

Arguments

b: width of the plate or half the width of the flange.

t: thickness of the flange.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case14 — Method

case14(d, t, E, F_y)

Classifies the compressive elements of a member subject to flexure.

Description of applicable member: Webs of doubly-symmetric I-shaped sections and channels.

Arguments

d: depth of the WT section

t_w: thickness of the web.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b

AISCSteel.ChapterBDesignRequirements.B4.TableB4⬝1a.case15 — Method

case15(b, t, E, F_y)

Classifies the compressive elements of a member subject to flexural.

Description of applicable member: Webs of doubly-symmetric I-shaped sections and channels.

Arguments

h:

For rolled sections: the clear distance between flanges less the fillet at each flange.

For built-up sections: clear distance between flanges when welds are used.

t_w: thickness of the web.

E: elastic section modulous.

F_y: yield strength of steel.

Returns

FlexureBucklingType enum whose value is either Compact, Noncompact, or Slender

Reference

AISC Table B4.1b